% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{add_pk}
\alias{add_pk}
\title{Add various `pkproto` objects to a `pk` object}
\usage{
add_pk(pk_obj, object, objectname)
}
\arguments{
\item{pk_obj}{The `pk` object}

\item{object}{The `pkproto` object to be added}

\item{objectname}{The name of the `pkproto` object to be added}
}
\value{
The `pk` object modified by the addition.
}
\description{
Add various `pkproto` objects to a `pk` object
}
