% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_1comp_fun.R
\name{analytic_1comp_fun}
\alias{analytic_1comp_fun}
\title{Analytical 1-compartment function}
\usage{
analytic_1comp_fun(params, dose, times, time.units = "h", iv.dose)
}
\arguments{
\item{params}{A named list of model parameter values. Must include:
\describe{
\item{kelim}{Elimination rate, 1/h}
\item{Vdist}{Volume of distribution, L/kg body weight}}
For oral administration (\code{iv.dose} FALSE), \code{params} must also include:
\describe{
\item{Fgutabs}{Oral bioavailability, unitless fraction}
\item{kgutabs}{Oral absorption rate, 1/h}}}

\item{dose}{Dose in mg/kg}

\item{times}{A vector of observation times in hours}

\item{time.units}{Either (h)ours or (d)ays (defaults to hours)}

\item{iv.dose}{TRUE for single IV bolus dose; FALSE for single oral dose}
}
\value{
A matrix of \code{time}, plasma concentration, and area under the curve (AUC).
}
\description{
Analytical 1-compartment function
}
\author{
Caroline Ring
}
