% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzepkdata.R
\name{analyze.pk.data}
\alias{analyze.pk.data}
\title{Actually does the fitting}
\usage{
analyze.pk.data(
  fitdata,
  this.cas,
  paramnames,
  modelfun,
  model,
  this.reference = NULL
)
}
\arguments{
\item{fitdata}{A data.table of concentration vs. time data for a given
chemical}

\item{this.cas}{A CAS number for the chemical to be fitted}

\item{paramnames}{A list of names of the model parameters to be fitted}

\item{modelfun}{"analytic" to use the analytic model solution, "full" to use
the full ODE model}

\item{model}{The model to fit, either "1compartment" or "2compartment"
(other models not implemented for now)}
}
\value{
A single row of fitted parameter values (arithmetic means, geometric
means, modes, arithmetic standard deviations, and geometric standard
deviations)
}
\description{
Fits model parameters to concentration vs. time data for a given chemical
}
\author{
Caroline Ring, John Wambaugh
}
