% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_pk_data.R
\name{analyze_pk_data}
\alias{analyze_pk_data}
\title{Fits a model to a set of data}
\usage{
analyze_pk_data(
  fitdata,
  this.dtxsid,
  paramnames,
  modelfun,
  model,
  this.reference = NULL,
  suppress.messages = FALSE,
  sig.figs = 5
)
}
\arguments{
\item{fitdata}{A data.table of concentration vs. time data for a given
chemical}

\item{this.dtxsid}{A DSSTox Substance Identifier for the chemical to be fitted}

\item{paramnames}{A list of names of the model parameters to be fitted}

\item{modelfun}{"analytic" to use the analytic model solution, "full" to use
the full ODE model}

\item{model}{The model to fit, either "1compartment" or "2compartment"
(other models not implemented for now)}

\item{this.reference}{placeholder}
}
\value{
A single row of fitted parameter values (arithmetic means, geometric
means, modes, arithmetic standard deviations, and geometric standard
deviations)
}
\description{
Fits one set of model parameters to one set of concentration vs. time data
}
\author{
Caroline Ring, John Wambaugh
}
