% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_subset.R
\name{analyze_subset}
\alias{analyze_subset}
\title{Fit a PK model to one set of concentration vs. time data}
\usage{
analyze_subset(
  fitdata,
  model,
  modelfun,
  sigma_ref = TRUE,
  fun_start = get_starts,
  fun_lower = get_lower_bounds,
  fun_upper = get_upper_bounds,
  fun_start_args = NULL,
  fun_lower_args = NULL,
  fun_upper_args = NULL,
  suppress.messages = FALSE,
  sig.figs = 5,
  factr = 1e+07
)
}
\arguments{
\item{fitdata}{A \code{data.frame} containing the set of concentration vs.
time data to be fitted. See Details for expected variables. See also
\code{\link{preprocess_data}}.}

\item{model}{Which general model should be fit for each chemical. Presently,
only "1compartment" and "2compartment" are implemented.}

\item{modelfun}{Either "analytic" or "full" -- whether to fit using the
analytic solution to the model, or the full ODE model. Presently,
"analytic" is recommended (because the analytic solution is exact and much
faster).}
}
\description{
Fit a specified PK model to one set of concentration vs. time data and return
a set of fitted parameter values.
}
\details{
# Expected variables in \code{fitdata}

\describe{\item{\code{Time}}{Time for each data point}
\item{\code{Value}}{Concentration for each data point}
\item{\code{Dose}}{Dose for each data point} \item{\code{DTXSID}}{DSSTox
Substance ID identifying the substance being dosed. This function expects
only one unique value in this column; if there are multiple values, it will
stop with an error.} \item{\code{Species}}{String identifying the species
being dosed. This function expects only one unique value in this column; if
there are multiple values, it will stop with an error.}
\item{\code{Reference}}{String identifying the reference or study for each
data point. Each reference is assumed to have its own residual error standard
deviation.} }
}
