% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starts.R
\name{assign_start}
\alias{assign_start}
\title{Assign starting value}
\usage{
assign_start(param_name, param_value, msg, par_DF, start_from)
}
\arguments{
\item{param_name}{Character: The name of the parameter whose starting value
should be checked and updated}

\item{param_value}{Numeric: The proposed starting value for the parameter}

\item{msg}{Character: The proposed message to use if the proposed starting
value is acepted}

\item{par_DF}{Data frame: A `data.frame` of parameters and bounds, as
produced by [get_opt_params], [get_lower_bounds], and [get_upper_bounds].
Must have row names corresponding to parameter names, and must have
variables `lower_bound` and `upper_bound`.}

\item{start_from}{Character: A vector of parameter names for which to
assign starting values from data. May be "all" to assign all from data, or
NULL, NA, character(0), or `''` to assign none from data.}
}
\value{
If the proposed starting value is within the bounds for the specified
  parameter, this function returns `par_DF` updated with variables
  `start_value` and `start_value_msg` in the row for the specified parameter,
  now containing `param_value` and `msg`, respectively. If those variables did
  not already exist in `par_DF`, they will be added. If the proposed starting
  value is *not* within bounds, this function returns `par_DF` unchanged.
}
\description{
Utility function to assign starting value for a parameter, if within
specified bounds
}
\author{
Caroline Ring
}
