% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_1comp.R
\name{auc_1comp}
\alias{auc_1comp}
\title{Analytic AUC for 1-compartment model}
\usage{
auc_1comp(time, params, dose, iv.dose)
}
\arguments{
\item{time}{A vector of times in hours}

\item{params}{A named list of model parameter values. Must include:
\describe{
\item{kelim}{Elimination rate, 1/h}
\item{Vdist}{Volume of distribution, L/kg body weight}}
For oral administration (\code{iv.dose} FALSE), \code{params} must also include:
\describe{
\item{Fgutabs}{Oral bioavailability, unitless fraction}
\item{kgutabs}{Oral absorption rate, 1/h}}}

\item{dose}{Dose in mg/kg}

\item{iv.dose}{TRUE for single IV bolus dose; FALSE for single oral dose}
}
\value{
A vector of plasma AUC values, evaluated at each time point in `time`.
}
\description{
Calculate area under the plasma concentration vs. time curve for the
1-compartment model, using an analytical equation (the integral of the
1-compartment model equation with respect to time).
}
\details{
# Model equations

## IV dosing

\deqn{\frac{D}{V_d k_e}  (1 - \exp(-t k_e))}

## Oral dosing

\deqn{\frac{dose Fa ka}{V_d (k_a-k_e)} ((\frac{1}{k_e} - \frac{1}{k_a}) +
(\frac{-\exp(-t k_e)}{k_e} + \frac{\exp(-t k_a)}{k_a} )}
}
\author{
Caroline Ring, John Wambaugh
}
