% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_2comp.R
\name{auc_2comp}
\alias{auc_2comp}
\title{Analytical AUC for the 2-compartment model}
\usage{
auc_2comp(params, time, dose, iv.dose)
}
\arguments{
\item{params}{A named list of parameter values including the following:
\describe{ \item{k12}{Rate at which compound moves from central to
peripheral compartment} \item{k21}{Rate at which compound moves from
peripheral to central compartment} \item{kelim}{Elimination rate}
\item{V1}{Apparent volume of central compartment} } For oral administration
(\code{iv.dose} FALSE), \code{params} must also include: \describe{
\item{Fgutabs}{Oral bioavailability} \item{kgutabs}{Rate of absorption from
gut} }}

\item{time}{A vector of time values, in hours}

\item{dose}{A dose in mg/kg}

\item{iv.dose}{TRUE for single IV bolus dose, FALSE for single oral dose}
}
\value{
A vector of plasma AUC values, evaluated at each time point in `time`.
}
\description{
Calculate area under the plasma concentration vs. time curve for the
1-compartment model, using an analytical equation (the integral of the
1-compartment model equation with respect to time).
}
\author{
Caroline Ring, John Wambaugh
}
