% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_flat.R
\name{auc_flat}
\alias{auc_flat}
\title{AUC for flat model}
\usage{
auc_flat(params, time, dose, iv.dose)
}
\arguments{
\item{params}{A named list of model parameter values. Must include:
\describe{
\item{A}{Average concentration}
}}

\item{time}{A vector of times in hours}

\item{dose}{Dose in mg/kg}

\item{iv.dose}{TRUE for single IV bolus dose; FALSE for single oral dose}
}
\value{
A vector of AUC values evaluated at each time point in \code{time}.
}
\description{
AUC for flat model
}
\author{
Caroline Ring, John Wambaugh, Chris Cook
}
