% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_resids.R
\name{bp_test}
\alias{bp_test}
\title{Breusch-Pagan test of heteroscedasticity
Evaluate heteroscedasticity using Breusch-Pagan test
The Breusch-Pagan test performs a linear regression of squared residuals on one or more explanatory variables. The test statistic is formed by the R-squared value of this regression multiplied by the number of residuals. This test statistic obeys a chi-squared distribution with S-1 degrees of freedom, where S is the number of explanatory terms (plus an intercept). The p-value of the test is given accordingly.}
\usage{
bp_test(varformula, data)
}
\arguments{
\item{varformula}{A formula for the variance linear model. The left-hand side of this formula must be the squared residuals.}

\item{data}{A `data.frame` containing all variables referenced in `varformula`.}
}
\value{
A list with the following elements:
- `statistic` : The test statistic
- `p.value`: The p-value of the test
- `r.squared`: The R-squared of the linear regression described in `varformula`
- `n`: The number of observations
- `df`: The chi-squared degrees of freedom
}
\description{
Breusch-Pagan test of heteroscedasticity
Evaluate heteroscedasticity using Breusch-Pagan test
The Breusch-Pagan test performs a linear regression of squared residuals on one or more explanatory variables. The test statistic is formed by the R-squared value of this regression multiplied by the number of residuals. This test statistic obeys a chi-squared distribution with S-1 degrees of freedom, where S is the number of explanatory terms (plus an intercept). The p-value of the test is given accordingly.
}
\references{
https://rpubs.com/cyobero/187387
}
\author{
Caroline Ring
}
