% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rmse.R
\name{calc_rsq}
\alias{calc_rsq}
\title{Calculate r-squared for observed vs. predicted values}
\usage{
calc_rsq(group_mean, group_sd, group_N, pred, log = FALSE)
}
\arguments{
\item{group_mean}{Numeric vector: Observed sample means for summary data, or
observed values for non-summary data. Censored observations should *not* be
NA; they should be substituted with some value at or below the
corresponding LOQ. Even if `log %in% TRUE`, these should *not* be
log-transformed.}

\item{group_sd}{Numeric vector: Observed sample SDs for summary data. For
non-summary data (individual-subject observations), the corresponding
element of `group_sd` should be set to 0. Even if `log %in% TRUE`, these
should *not* be log-transformed.}

\item{group_N}{Numeric vector: Observed sample number of subjects for summary
data. For non-summary data (individual-subject observations), `group_n`
should be set to 1.}

\item{pred}{Numeric vector: Model-predicted value corresponding to each
observed value. Even if `log %in% TRUE`, these should *not* be
log-transformed.}

\item{log}{Logical. FALSE (default) means that RMSE is computed for
natural-scale observations and predictions. TRUE means that observations
and predictions will be log-transformed before RMSE is calculated (see
Details).}

\item{group_LOQ}{Numeric vector: Reported limit of quantification (LOQ) (i.e,
left-censoring limit) for each observation. May contain `NA_real_`s for
non-censored observations. Even if `log %in% TRUE`, these should *not* be
log-transformed.}
}
\value{
A numeric scalar: the R-squared value for observations vs.
  predictions.
}
\description{
Calculate the square of the Pearson correlation coefficient (r) between observed
and model-predicted values
}
\details{
Calculate the square of the Pearson correlation coefficient (r) between observed
and model-predicted values, when observed data may be left-censored (non-detect) or may be
reported in summary form (as sample mean, sample standard deviation, and
sample number of subjects). Additionally, handle the situation when observed
data and predictions need to be log-transformed before RMSE is calculated.

\eqn{r^2} is calculated according to the following
formula, to properly handle observations reported in summary format:

\deqn{ r^2 = \left(
\frac{
\sum_{i=1}^G \mu_i n_i \bar{y}_i -
(\bar{\mu} + \bar{y}) \sum_{i=1}^G n_i \mu_i +
(\bar{mu} \bar{y}) \sum_{i=1}^G n_i
}
{ \sqrt{
\sum_{i=1}^G (n_i - 1) s_i^2 +
\sum_{i=1}^G n_i \bar{y}_i^2 -
2 \bar{y} \sum_{i=1}^G n_i \bar{y}_i +
N + \bar{y}^2
  }
\sqrt{
\sum_{i=1}^G n_i \mu_i^2 -
2 \bar{y} \sum_{i=1}^G n_i \mu_i +
N + \bar{y}^2
}
} \right)^2
}

In this formula, there are \eqn{G} groups (reported observations). \eqn{n_i}
is the number of subjects for group \eqn{i}. \eqn{\bar{y}_i} is the sample
mean for group \eqn{i}. \eqn{s_i} is the sample standard deviation for group
\eqn{i}.\eqn{\mu_i} is the model-predicted value for group \eqn{i}.
\eqn{\bar{y}} is the grand mean of observations:

\deqn{ \bar{y} = \frac{ \sum_{i=1}^G n_i \bar{y}_i }{\sum_{i=1}^G n_i} }

\eqn{\bar{\mu}} is the grand mean of predictions:

\deqn{ \bar{\mu} = \frac{ \sum_{i=1}^G n_i \mu_i }{\sum_{i=1}^G n_i} }

\eqn{N} is the grand total of subjects:

\deqn{N = \sum_{i=1}^G n_i}

For the non-summary case (\eqn{N} single-subject observations, with all
\eqn{n_i = 1}, \eqn{s_i = 0}, and \eqn{\bar{y}_i = y_i}), this formula
reduces to the familiar formula

\deqn{ r^2 = \left( \frac{\sum_{i=1}^N (y_i - \bar{y}) (\mu_i - \bar{\mu})}
{\sqrt{ \sum_{i=1}^N (y_i - \bar{y})^2 }
\sqrt{ \sum_{i=1}^N (\mu_i - \bar{\mu})^2 }
 } \right)^2
 }

# Left-censored data

If the observed value is censored, and the predicted value is less than the
reported LOQ, then the observed value is (temporarily) set equal to the
predicted value, for an effective error of zero.

If the observed value is censored, and the predicted value is greater than
the reported LOQ, the the observed value is set equal to the reported LOQ.

# Log transformation

If `log %in% TRUE`, then both the observed and predicted values will be
(natural) log-transformed before calculating the RMSE. In the case where
observed values are reported in summary format, each sample mean and sample
SD (reported on the natural scale, i.e. the mean and SD of natural-scale
individual observations) are used to produce an estimate of the log-scale
sample mean and sample SD (i.e., the mean and SD of log-transformed
individual observations), using [convert_summary_to_log()].

The formulas are as follows. Again, \eqn{\bar{y}_i} is the sample mean for
group \eqn{i}. \eqn{s_i} is the sample standard deviation for group \eqn{i}.

\deqn{\textrm{log-scale sample mean}_i = \log
\left(\frac{\bar{y}_i^2}{\sqrt{\bar{y}_i^2 + s_i^2}} \right)}

\deqn{\textrm{log-scale sample SD}_i = \sqrt{\log \left(1 +
\frac{s_i^2}{\bar{y}_i^2} \right)}}
}
\author{
Caroline Ring
}
