% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{coef.pk}
\alias{coef.pk}
\title{Get coefficients}
\usage{
\method{coef}{pk}(obj, model = NULL, method = NULL)
}
\arguments{
\item{obj}{A [pk] object}

\item{model}{Optional: Specify one or more of the fitted models whose
coefficients to return. If NULL (the default), coefficients will be returned for all of
the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
whose coefficients to return. If NULL (the default), coefficients will be returned for
all of the models in `obj$optimx_settings$method`.}
}
\value{
A named list of numeric matrixes. There is one list element named for
  each model in `model`. Each list element is a matrix with as many
  rows as items in `method`. The row names are the method names. The matrix column names are
  the names of the fitted parameters, including any error standard deviation
  hyperparameters (whose names begin with "sigma").
}
\description{
Extract coefficients from a fitted `pk` object
}
\author{
Caroline Ring
}
