% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rmse.R
\name{combined_sd}
\alias{combined_sd}
\title{Combined standard deviation}
\usage{
combined_sd(
  group_mean,
  group_sd,
  group_N,
  unbiased = TRUE,
  na.rm = TRUE,
  log = FALSE
)
}
\arguments{
\item{group_mean}{A numeric vector of group means.}

\item{group_sd}{A numeric vector of group SDs.}

\item{group_N}{A numeric vector of group sizes.}

\item{unbiased}{Logical. If TRUE, then `group_sd` is assumed to be the
unbiased estimator of population standard deviation (i.e. using `n-1` in
the denominator -- the way that `stats::sd()` calculates it), and the
returned combined SD is also the unbiased estimator of the combined
population SD. If FALSE, then `group_sd` is assumed to be the biased
estimator (using `n` in the denominator), and the returned value is also
the biased estimator of the combined population SD.}

\item{na.rm}{Logical. If TRUE (default), then any groups where mean, SD, *or* N
were NA will be dropped. If FALSE, they will be retained (and the result will
be NA).}
}
\value{
Numeric: the standard deviation of the combined population (i.e. if
  all the groups were concatenated into one large group).
}
\description{
Given mean, standard deviation, and N for some set of groups, calculate the
combined standard deviation. Note that the groups may not overlap.
}
\author{
Caroline Ring
}
