% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rmse.R
\name{convert_summary_to_log}
\alias{convert_summary_to_log}
\title{Convert sample mean and SD to log-scale}
\usage{
convert_summary_to_log(sample_mean, sample_SD)
}
\arguments{
\item{sample_mean}{Numeric: one or more sample means}

\item{sample_SD}{Numeric: one or more sample SDs}
}
\value{
A list with two named elements: "logmean" and "logSD", the log-scale
  sample means and log-scale sample SDs, respectively.
}
\description{
Estimate log-scale sample mean and standard deviation from natural-scale
sample mean and standard deviation
}
\details{
\eqn{\bar{y}_i} is the natural-scale sample mean for group \eqn{i}. \eqn{s_i}
is the natural-scale sample standard deviation for group \eqn{i}.

\deqn{\textrm{log-scale sample mean}_i = \log
\left(\frac{\bar{y}_i^2}{\sqrt{\bar{y}_i^2 + s_i^2}} \right)}

\deqn{
\textrm{log-scale sample SD}_i =
 \sqrt{
 \log \left(1 + \frac{s_i^2}{\bar{y}_i^2} \right)
 }
 }
}
\author{
Caroline Ring
}
