% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_1comp.R
\name{cp_1comp}
\alias{cp_1comp}
\title{Analytical 1-compartment model}
\usage{
cp_1comp(params, time, dose, iv.dose)
}
\arguments{
\item{params}{A named list of numeric parameter values. See Details for
requirements.}

\item{time}{A numeric vector of times in hours, reflecting the time point when
concentration is measured after the corresponding single bolus dose. Must be
same length as `dose` and `iv.dose`, or length 1.}

\item{dose}{A numeric vector of doses in mg/kg, reflecting single bolus doses
administered at time 0. Must be same length as `time` and `iv.dose`, or
length 1.}

\item{iv.dose}{A logical vector, reflecting the route of administration of
each single bolus dose. TRUE for single IV bolus dose; FALSE for single oral
bolus dose.  Must be same length as `time` and `dose`, or length 1.}
}
\value{
A vector of plasma concentration values (mg/L) corresponding to
 `time`.
}
\description{
Calculates plasma concentrations vs. time according to the analytical solution
for the 1-compartment model, for single bolus doses (IV and/or oral).
}
\details{
# Required parameters

`params` must include the following named items:
  \describe{
  \item{kelim}{Elimination rate, 1/h.}
  \item{Vdist}{Apparent volume of central compartment, L/kg BW. Or see below for
  `Fgutabs_Vdist`}
  }

For oral administration (if any `iv.dose == FALSE`), `params` must also
include:
  \describe{
  \item{Fgutabs}{Oral bioavailability, unitless fraction. Or see below for
  `Fgutabs_Vdist`}
  \item{kgutabs}{Rate of absorption from gut, 1/h.}
  }

For oral administration, in lieu of `Vdist` and `Fgutabs`, you may instead
provide `Fgutabs_Vdist`, the ratio of Fgutabs to Vdist (1/L). This is an
alternate parameterization for situations where `Fgutabs` and `Vdist` are not
identifiable separately (i.e., when oral TK data are available, but IV data
are not). If `Fgutabs` and `Vdist` are provided, they will override any value
provided for `Fgutabs_Vdist`.

If both oral and IV administration are specified (i.e., some `iv.dose == TRUE`
and some `iv.dose == FALSE`), then `Vdist` is required along with either
`Fgutabs` or `Fgutabs_Vdist`. (If `Vdist` and `Fgutabs_Vdist` are provided,
but `Fgutabs` is not provided, then `Fgutabs` will be calculated from `Vdist`
and `Fgutabs_V1`.)
}
\author{
Caroline Ring, John Wambaugh
}
