% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_2comp.R
\name{cp_2comp}
\alias{cp_2comp}
\title{Analytical 2-compartment model}
\usage{
cp_2comp(params, time, dose, iv.dose)
}
\arguments{
\item{params}{A named list of parameter values including the following:
  \describe{
  \item{k12}{Rate at which compound moves from central to peripheral
  compartment, 1/h.}
  \item{k21}{Rate at which compound moves from peripheral to central
  compartment, 1/h.}
  \item{kelim}{Elimination rate, 1/h.}
  \item{V1}{Apparent volume of central compartment, L/kg BW. Or see below for
  "Fgutabs_V1"}
  }

  For oral administration (\code{iv.dose} FALSE), \code{params} must also
  include:
  \describe{
  \item{Fgutabs}{Oral bioavailability, unitless fraction. Or see below for
  "Fgutabs_V1"}
  \item{kgutabs}{Rate of absorption from gut, 1/h.}
  }

  For oral administration, in lieu of "V1" and "Fgutabs", you may instead
  provide "Fgutabs_V1", the ratio of Fgutabs to V1 (1/L). This is an
  alternate parameterization for situations where "Fgutabs" and "V1" are not
  identifiable separately (i.e. when oral data are available, but IV data are
  not). If "Fgutabs" and "V1" are provided, then "Fgutabs_V1" will not be
  used.}

\item{time}{A numeric vector of time values, in hours}

\item{dose}{A numeric vector of doses in mg/kg}

\item{iv.dose}{A logical vector: TRUE for single IV bolus dose, FALSE for single oral dose}
}
\value{
A vector of plasma concentration values (mg/L) corresponding to each value
  in \code{time}
}
\description{
Calculates plasma concentration according to the analytical solution for the
2-compartment model.
}
\author{
Caroline Ring, John Wambaugh
}
