% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_conc_sd.R
\name{estimate_conc_sd}
\alias{estimate_conc_sd}
\title{Impute missing sample standard deviations}
\usage{
estimate_conc_sd(
  dat,
  reference_col = "Reference",
  chem_col = "DTXSID",
  media_col = "Media",
  species_col = "Species",
  value_col = "Value",
  sd_col = "Value_SD",
  n_subj_col = "N_Subjects"
)
}
\arguments{
\item{dat}{A `data.frame` containing the concentration vs. time data to be
imputed. For example, from CvTdb.}

\item{reference_col}{Name of the variable in `dat` identifying unique
references. Default "Reference".}

\item{chem_col}{Name of the variable in `dat` identifying unique
chemicals/substances. Default "DTXSID."}

\item{media_col}{Name of the variable in `dat` identifying concentration
media (blood or plasma). Default "Media".}

\item{species_col}{Name of the variable in `dat` identifying species. Default
"Species".}

\item{value_col}{Name of the variable in `dat` that contains concentration
values (sample means). Default "Value".}

\item{sd_col}{Name of the variable in `dat` that contains concentration
sample standard deviation values. If this variable does not already exist
in `dat`, it will be created and filled with `NA_real_`. Default
"Value_SD".}

\item{n_subj_col}{Name of the variable in `dat` that contains the number of
subjects for each sample mean/SD. Default "N_Subjects".}
}
\value{
The same `data.frame` as `dat`, but with missing values in the
  variable `sd_col` imputed as described in Details.
}
\description{
For multi-subject summary data reported as sample mean, sample standard
deviation (SD), and number N of subjects, impute values for any missing
sample SDs.
}
\details{
This function imputes missing SDs as the minimum non-missing SD per
reference, chemical, species, and medium. The idea is that a missing SD
probably occurred because there were no visible error bars on a point on a
graph, so data extractors/curators reported the SD missing. In this case, the
effective errorbars would be the size of the symbol itself. The minimum
non-missing SD is a conservative estimate of the symbol size; if error bars
were just barely visible, this would be the minimum non-missing SD.

If there are no non-missing SDs for a given reference, chemical, species, and
medium, then this function imputes missing sample SDs as equal to the sample
means. This assumption was made after observation of a plot of sample SD vs.
sample mean for multi-subject data across chemicals and studies. The relation
is effectively on the identity line.

If there are no non-missing SDs for a given reference, chemical, species, and
medium, and a sample mean is NA (missing or below LOD), then a corresponding
missing sample SD will remain missing.
}
\author{
Caroline Ring
}
