% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_loq.R
\name{estimate_loq}
\alias{estimate_loq}
\title{Assign an estimated limit of quantification wherever missing}
\usage{
estimate_loq(
  cvt.data,
  source.col = "fk_study_id",
  chem.col = "dsstox_casrn",
  media.col = "conc_medium_normalized",
  value.col = "conc",
  calc.loq.col = "calc_loq"
)
}
\arguments{
\item{cvt.data}{A table of concentration vs. time data with at least the
following columns:}

\item{source.col}{The column indicating different research studies (and
there for different mass spectrometry methods)}

\item{chem.col}{The column indicating chemical identity}

\item{media.col}{The column indicating sample medium}

\item{value.col}{The column indicating concentration}

\item{calc.loq.col}{The column to be added to the table holding the estimated
LOQ's}
}
\value{
A table with column "calc_loq" added
}
\description{
Sometimes in mass spectrometry a compound is detected, but the signal is
too low and falls outside the range that can be interpreted quantitatively
(that is, as a concentration). These detects can still be used to refine
parameter estimates by preferring curve fits that predict concentrations
below the known "limit of quantitation" (LOQ) -- the lowest detectable quantifiable
concentration. LOQ's are not always available from data reported in the
literature, so this function estimates a LOQ such that all the reported
values are above the LOQ. The LOQ is assumed to vary from study to study,
as each study would have different mass spectrometry equipment and methods.
LOQ also is assumed to vary by media, as preparation of samples may be different.
}
\author{
John Wambaugh
}
