% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_fit.R
\name{evaluate_fit}
\alias{evaluate_fit}
\title{Evaluate goodness of fit}
\usage{
evaluate_fit(
  cvt_pre,
  fit_flat,
  fit_1comp,
  fit_2comp,
  fit_conc_dose = TRUE,
  optimx_args = list(method = "bobyqa", itnmax = 1e+06, control = list(kkt = FALSE))
)
}
\arguments{
\item{cvt_pre}{Original concentration vs. time data, pre-processed using
`preprocess_data()`.}

\item{fit_flat}{Fitting output for the flat model: the output of `fit_all()`
with `model = "flat"`}

\item{fit_1comp}{Fitting output for the 1-compartment model: the output of
`fit_all()` with `model = "1compartment"`}

\item{fit_2comp}{Fitting output for the 2-compartment model: the output of
`fit_all()` with `model = "2compartment"`}

\item{fit_dose_norm}{Logical: Whether the model fits were done on
dose-normalized data (TRUE), or not (FALSE). This needs to be the same as
whatever you used when calling `fit_all()`, or the goodness-of-fit metrics
will be incorrect!}
}
\value{
A `data.table` of goodness-of-fit measures for each dataset, each
  model, and each analysis (joint, separate, and pooled)
}
\description{
Compute multiple metrics to evaluate how well the various TK models fit the
data.
}
\author{
Caroline Ring
}
