% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_resids.R
\name{evaluate_resids}
\alias{evaluate_resids}
\title{Evaluate goodness of fit based on residuals}
\usage{
evaluate_resids(
  cvt_pre,
  fit_flat,
  fit_1comp,
  fit_2comp,
  fit_conc_dose = TRUE,
  fit_log_conc = FALSE,
  optimx_args = list(method = "bobyqa", itnmax = 1e+06, control = list(kkt = FALSE))
)
}
\arguments{
\item{cvt_pre}{Original concentration vs. time data, pre-processed using
`preprocess_data()`.}

\item{fit_flat}{Fitting output for the flat model: the output of `fit_all()`
with `model = "flat"`}

\item{fit_1comp}{Fitting output for the 1-compartment model: the output of
`fit_all()` with `model = "1compartment"`}

\item{fit_2comp}{Fitting output for the 2-compartment model: the output of
`fit_all()` with `model = "2compartment"`}

\item{fit_conc_dose}{Logical: Whether the model fits were done on
dose-normalized data (TRUE), or not (FALSE). This needs to be the same as
whatever you used when calling `fit_all()`, or the goodness-of-fit metrics
will be incorrect!}

\item{fit_log_conc}{Logical: Whether the model fits were done on
log-transformed data (TRUE), or not (FALSE). This needs to be the same as
whatever you used when calling `fit_all()`, or the goodness-of-fit metrics
will be incorrect!}
}
\value{
A `data.table` of goodness-of-fit measures for each dataset, each
  model, and each analysis (joint, separate, and pooled)
}
\description{
Evaluate multiple metrics of goodness of fit based on residuals
}
\details{
For each dataset (chemical and species) and each fitted model, this function
first calculates the predictions and residuals, then evaluates the following:

- RMSE (root mean squared error)
- R-squared of observations vs. predictions
- The p-value from the Breusch-Pagan test of heteroscedasticity
- Estimated "overall" error standard deviation (details below)

Unique fitted models are defined by unique combination of analysis type
("Joint", "Separate", or "Pooled"), studies analyzed, and model ("flat",
"1compartment", or "2compartment").

# Root mean squared error (RMSE)

RMSE is calculated in [calc_rmse()].

# Estimated overall error standard deviation

The "overall" error standard deviation is the value for residual error
standard deviation ("sigma") that minimizes the log-likelihood function when
the TK model parameters are fixed at their fitted values. This is essentially
a more rigorous RMSE, because it correctly models the likelihood of
non-detect (below-LOQ) observations as the cumulative probability that the
model prediction is below the reported LOQ, rather than making any
substitutions for non-detects.
}
\author{
Caroline Ring
}
