% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all.R
\name{fit_all}
\alias{fit_all}
\title{Main fitting function}
\usage{
fit_all(data.set, model, modelfun = NA, ratio.data.to.dose = 1)
}
\arguments{
\item{data.set}{A table of concentration-time data. Preferably
\code{pkdataset_nheerlcleaned} or \code{pkdataset_nheerlorig}.}

\item{model}{Which general model should be fit for each chemical. Presently,
only "1compartment" and "2compartment" are implemented.}

\item{modelfun}{Either "analytic" or "full" -- whether to fit using the
  analytic solution to the model, or the full ODE model. Presently,
  "analytic" is recommended (because the analytic solution is exact and much
  faster).
#param ratio between the weights used to report the data and the weights used 
  for the dose. For example, ug/L data and mg/kg/day dose would be 0.001
  (defaults to 1)}
}
\value{
A data.table of fitted parameter values for each chemical.
}
\description{
Fits parameters of a specified model to concentration-time data given in
data.set
}
\author{
Caroline Ring, John Wambaugh
}
