% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfun.R
\name{fitfun}
\alias{fitfun}
\title{Compute model-predicted concentration values for a vector of time values}
\usage{
fitfun(design.times, design.dose, design.iv, design.times.max, design.time.step,
  modelfun, model, model.params)
}
\arguments{
\item{design.times}{A vector of times from the design matrix}

\item{design.dose}{A dose from the design matrix}

\item{design.iv}{TRUE if IV data, FALSE if PO data}

\item{design.times.max}{The maximum of design.times}

\item{design.time.step}{The maximum number of time steps per hour.}

\item{modelfun}{"analytic" to use analytic model solution, "full" to use full
ODE model}

\item{model}{"1compartment" or "2compartment"}

\item{model.params}{A named vector of model parameter values.}
}
\value{
A vector of the model-predicted plasma concentration values of the
 same length as design.times, for each time in design.times
}
\description{
Compute model-predicted concentration values for a vector of time values
}

