% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_loq.R
\name{fix_loq}
\alias{fix_loq}
\title{Helper function to calculate LOQ for processed data}
\usage{
fix_loq(data)
}
\arguments{
\item{data}{data.table with Value and LOQ column names}
}
\value{
A data.table with assigned LOQ values for those that were NA
}
\description{
Helper function to calculate LOQ for processed data
}
\author{
Christopher Cook
}
