% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_predictions.R
\name{get_all_predictions}
\alias{get_all_predictions}
\title{Make predictions for all model fits}
\usage{
get_all_predictions(cvt_pre, fit_flat, fit_1comp, fit_2comp)
}
\arguments{
\item{cvt_pre}{Original concentration vs. time data, pre-processed using
[preprocess_data()].}

\item{fit_flat}{Fitting output for the flat model: the output of [fit_all()]
with `model = "flat"`}

\item{fit_1comp}{Fitting output for the 1-compartment model: the output of
[fit_all()] with `model = "1compartment"`}

\item{fit_2comp}{Fitting output for the 2-compartment model: the output of
[fit_all()] with `model = "2compartment"`}
}
\value{
A `data.table` with `nrow(cvt_pre) * 9` rows and `ncol(cvt_pre) + 5`
  variables. It contains model predictions for each observation in `cvt_pre`,
  for each analysis type ("Joint", "Separate", and "Pooled") and each model
  ("flat", "1compartment", and "2compartment"). It has the same variables as
  `cvt_pre`, plus additional variables `Analysis_Type`, `Studies.Analyzed`,
  `model`, `pred_conc`, and `pred_auc`. `Analysis_Type`, `Studies.Analyzed`,
  and `model` specify which analysis type, which set of studies, and which
  model were used to produce the model fit for which predictions are made.
  `pred_conc` contains the predicted concentration (in mg/L), and `pred_auc`
  contains the predicted cumulative AUC (area under the concentration-time
  curve evaluated from time 0 to the time in the corresponding row, units of
  mg/L*hour), for the dose, time, route, and medium corresponding to the
  observation in each row.
}
\description{
Predict concentrations for all chemicals, species, analysis types, and models
}
\author{
Caroline Ring
}
