% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_sd.R
\name{get_error_sd}
\alias{get_error_sd}
\title{Estimate pooled error standard deviation with fixed parameters}
\usage{
get_error_sd(
  pk_fit_row,
  newdata,
  model,
  modelfun = "analytic",
  log_trans = TRUE,
  dose_norm = FALSE,
  optimx_args = list(method = "bobyqa", itnmax = 1e+06, control = list(kkt = FALSE))
)
}
\arguments{
\item{pk_fit_row}{One row from a table of fitted PK parameters, as produced by
[merge_fits()]. Must contain variables `DTXSID`, `Species`,
`Analysis_Type`, and `Studies.Analyzed`. Must also contain variables
corresponding to the parameters of the model specified in argument
`model`, as required for the route and media of data in `newdata` (as
given by [get_model_paramnames()]). These variables must be named as
`[param].[model]`. For example, for the 1-compartment model, if `newdata`
contains only IV-dosing data measured in plasma, then `pk_fit` must contain
variables named `kelim.1compartment` and `Vdist.1compartment`.}

\item{newdata}{A `data.table` which must contain variables named `Time`,
`Dose`, `iv`, and `Media`.}

\item{model}{The model to evaluate: one of 'flat', '1compartment', or
'2compartment'}

\item{optimx_args}{A named list of additional arguments to [optimx::optimx()],
 other than `par`, `fn`, `lower`, and `upper`. Default is:

   ```
    list(
          "method" = "bobyqa",
          "itnmax" = 1e6,
         "control" = list("kkt" = FALSE))
   ```
 Briefly:  `"method"` allows you to select an optimization algorithm.
 `"itnmax"` controls the maximum number of iterations allowed in an attempt
 to optimize. `"control"` is itself a named list of control parameters
 relevant to the selected method. See documentation for [optimx::optimx()]
 for more details and more options. Note lower and upper bounds (box
 constraints) will be supplied; if you want them to be respected, please
 choose a method that allows box constraints (e.g. "bobyqa" or "L-BFGS-B").}
}
\value{
A one-row `data.table` with a variable `sigma` reporting the estimated
 residual error standard deviation, and other variables: `fevals` reporting
 the number of fucntion evaluations performed as returned by
 [optimx::optimx()]; `convcode` reporting the integer convergence code
 returned by [optimx::optimx()]; `niter`reporting the number of iterations
 performed by [optimx::optimx()]; `method` reporting the optimization method
 used (as specified in input argument `optimx_args`); and variables reporting
 each control parameter specified in `optimx_args`, prefixed with `control_`.
 (For example, using the default value of `optimx_args`, there will be a
 variable named `control_kkt` whose value will be FALSE.)
}
\description{
Given a table of model parameters (e.g. fitted using [fit_all()]), estimate
the residual error standard deviation assuming residuals are iid following a
zero-mean normal distribution
}
\author{
Caroline Ring
}
