% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elbow.R
\name{get_middle}
\alias{get_middle}
\title{Helper function to get middle point}
\usage{
get_middle(x, y, ...)
}
\arguments{
\item{x}{A numeric vector of `x` values}

\item{y}{A numeric vector of `y` values}

\item{...}{Optional additional arguments to [stats::approx()] (other than
`x`, `y`, and `xout`). See documentation for [stats::approx()] for details
and options.}
}
\value{
A list with two named numeric scalar elements, `x` and `y`. `x`
  contains the middle `x` value from the input data. `y` contains the
  interpolated `y` value at that middle `x` value.
}
\description{
Given a set of data provided as `x` and `y` vectors, return the middle `x`
value and an approximation of the corresponding `y` value at that `x` value.
}
\details{
The middle `x` value is defined as `mean(range(x, na.rm = TRUE))`. To
approximate the corresponding `y` value, [stats::approx()] is used.

This helper function is called by [get_elbow()] in case
[akmedoids::elbow_point()] fails.
}
\author{
Caroline Ring
}
