% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_function.R
\name{get_model_function}
\alias{get_model_function}
\title{Get model function}
\usage{
get_model_function(model)
}
\arguments{
\item{model}{The name of a model. One of 'flat', '1compartment', or
'2compartment'.}
}
\value{
The function corresponding to `model`: `cp_flat()` for 'flat',
  `cp_1comp()` for '1compartment', `cp_2comp()` for '2compartment'. Returned
  as a function object, not as a character string.
}
\description{
Get the function that predicts concentration for a given model name
}
\author{
Caroline Ring
}
