% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_paramnames.R
\name{get_model_paramnames}
\alias{get_model_paramnames}
\title{Get model parameter names}
\usage{
get_model_paramnames(model)
}
\arguments{
\item{model}{The name of the model}
}
\description{
Get parameter names required by a specified model
}
