% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opt_params.R
\name{get_opt_params}
\alias{get_opt_params}
\title{Determine which parameters to fit (optimize)}
\usage{
get_opt_params(
  model,
  fitdata,
  param_names = NULL,
  sigma_ref = TRUE,
  suppress.messages = FALSE
)
}
\arguments{
\item{model}{Character: the name of the model to parameterize.}

\item{fitdata}{Data.frame: The set of concentration-dose-time data to be used
to fit the model parameters. Requires a variable named "Route" which
contains the dosing route, either "po" (oral dosing) or "iv" (intravenous
dosing).}

\item{param_names}{Optional: A character vector of parameter names needed by
the model. Default NULL to automatically determine these from `model` by
calling [get_model_paramnames].}
}
\value{
A named logical vector whose names are `param_names`, indicating
  whether to fit each model parameter (TRUE) or hold it constant (FALSE).
}
\description{
Determine whether to fit (optimize) each model parameter from the data or
hold it constant
}
\details{
All model parameters will be fitted (optimized) with the following
exceptions:

For 1-compartment and 2-compartment models:

If no oral dosing data are available, then the parameter "kgutabs" will not
be estimated from the data. Instead it will be held constant at its
"starting" value while other parameters are estimated.

If data do not include both oral dosing and IV dosing, then the parameter
"Fgutabs" will not be estimated from the data. Instead it will be held
constant at its "starting" value while other parameters are estimated.

If non-NULL `param_names` do not match the result of
`get_model_parameters(param_names)`, this function will throw a warning, but
will otherwise proceed using `param_names`.
}
