% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_2comp.R
\name{get_params_2comp}
\alias{get_params_2comp}
\title{Get parameters for 2-compartment model}
\usage{
get_params_2comp(
  data,
  lower_bound = c(kelim = 1e-04, V1 = 1e-04, k21 = 1e-04, k21 = 1e-04, Fgutabs = 1e-04,
    kgutabs = 1e-04, Fgutabs_V1 = 1e-04, Rblood2plasma = 1e-04),
  upper_bound = c(kelim = 1e+06, V1 = 1e+06, k21 = 1e+06, k12 = 1e+06, Fgutabs = 1,
    kgutabs = 1e+06, Fgutabs_v1 = 10000, Rblood2plasma = 1e+06),
  param_units = ggplot2::aes(kelim = paste0("1/", unique(Time_trans.Units)), V1 =
    paste0("(", unique(Dose.Units), ")", "/", "(", unique(Conc.Units), ")"), k21 =
    paste0("1/", unique(Time_trans.Units)), k12 = paste0("1/", unique(Time_trans.Units)),
    Fgutabs = "unitless fraction", kgutabs = paste0("1/", unique(Time_trans.Units)),
    Fgutabs_V1 = paste0("(", unique(Conc.Units), ")", "/", "(", unique(Dose.Units), ")"),
    Rblood2plasma = "unitless ratio")
)
}
\arguments{
\item{data}{The data set to be fitted, with harmonized variable names (e.g. the result of [preprocess_data()]).}
}
\value{
A `data.frame`with the following variables:
- `param_name`: Names of the model parameters
- `param_units`: Units of the model parameters
- `optimize_param`: TRUE if each parameter is to be estimated from the data; FALSE otherrwise
- `use_param`: TRUE if each parameter is to be used in evaluating the model; FALSE otherwise
}
\description{
Get parameters for 2-compartment model and determine whether each is to be
estimated from the data
}
\details{
The full set of model parameters for the 1-compartment model includes `V1`,
`kelim`, `k12`, `k21`, `kgutabs`, and `Fgutabs`.

## IV data, no oral data

If IV dosing data are available, but no oral dosing data are available, then
only the parameters `V1`, `kelim`, `k12`, and `k21` will be estimated from
the data. The parameters `kgutabs` and `Fgutabs` cannot be estimated from IV
data alone.

## Oral data, no IV data

If oral dosing data are available, but no IV dosing data are available, then
the parameters `kelim`, `k12`, `k21`, and `kgutabs` will be estimated from
the data. However, the parameters `Fgutabs` and `V1` cannot be identified
separately. From oral data alone, only the ratio `Fgutabs/V1` can be
identified. This ratio is represented by a single parameter named
`Fgutabs_V1`. `Fgutabs` and `V1` will not be optimized, but `Fgutabs_V1` will
be optimized, along with `kelim`, `k12`, `k21`, and `kgutabs`.

## Oral data and IV data

If both oral and IV dosing data are available, then `V1`, `kelim`, `k12`,
`k21`, `kgutabs`, and `Fgutabs` will all be estimated from the data.

# Blood and plasma data

If both blood and plasma data are available, then `Rblood2plasma` will be estimated from the data.

# Blood data, no plasma data

If blood data but no plasma data are available, then `Rblood2plasma` will be used in model evaluation, but held constant at 1, not estimated from the data.

# Plasma data, no blood data

If plasma data but no blood data are available, then `Rblood2plasma` will neither be estimated nor be used in model evaluation at all.
}
\author{
Caroline Ring
}
