% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_flat.R
\name{get_params_flat}
\alias{get_params_flat}
\title{Get parameters to be optimized for flat model}
\usage{
get_params_flat(data)
}
\arguments{
\item{data}{The data set to be fitted, with harmonized variable names (e.g. the result of [preprocess_data()]).}
}
\value{
A `data.frame`with the following variables:
- `param_names`: Names of the model parameters
- `param_units`: Units of the model parameters
- `opt_params`: TRUE if each parameter is to be estimated from the data; FALSE otherrwise
- `use_params`: TRUE if each parameter is to be used in evaluating the model; FALSE otherwise
}
\description{
The full set of model parameters for the flat model includes
`Vdist`, and `Fgutabs`. Whether each one can be estimated
from the data depends on what routes of administration are included in the
data.
}
\details{
## IV data, no oral data

If IV dosing data are available, but no oral dosing data are available, then
only the parameters `Vdist` and `kelim` will be estimated from the data. The
parameters `kgutabs` and `Fgutabs` cannot be estimated from IV data alone and will not be used to evaluate the model.

## Oral data, no IV data

If oral dosing data are available, but no IV dosing data are available, then
the parameters `kelim` and `kgutabs` can be estimated from the data. However,
the parameters `Fgutabs` and `Vdist` cannot be identified separately. From
oral data alone, only the ratio `Fgutabs/Vdist` can be identified. This ratio
is represented by a single parameter named `Fgutabs_Vdist`. `Fgutabs` and
`Vdist` will not be estimated nor used in model evaluation, but `Fgutabs_Vdist` will be estimated, along
with `kelim` and `kgutabs`.

## Oral data and IV data

If both oral and IV dosing data are available, then `Vdist`, `kelim`,
`kgutabs`, and `Fgutabs` will all be estimated from the data.

# Blood and plasma data

If both blood and plasma data are available, then `Rblood2plasma` will be estimated from the data.

# Blood data, no plasma data

If blood data but no plasma data are available, then `Rblood2plasma` will be used in model evaluation, but held constant at 1, not estimated from the data.

# Plasma data, no blood data

If plasma data but no blood data are available, then `Rblood2plasma` will neither be estimated nor be used in model evaluation at all.
}
\author{
Caroline Ring
}
