% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get model predicted concentration and AUC}
\usage{
get_predictions(
  pk_fit,
  newdata,
  DTXSID_in,
  Species_in,
  Analysis_Type_in,
  Studies.Analyzed_in,
  model_in
)
}
\arguments{
\item{pk_fit}{A table of fitted PK parameters, as produced by
`postprocess_data()`. Must contain variables `DTXSID`, `Species`,
`Analysis_Type`, and `Studies.Analyzed`. Must also contain variables
corresponding to the parameters of the model specified in argument
`model_in`, as required for the route and media of data in `newdata` (as
given by [get_model_paramnames()]). These variables must be named as
`[param].[model]`. For example, for the 1-compartment model, if `newdata`
contains only IV-dosing data measured in plasma, then `pk_fit` must contain
variables named `kelim.1compartment` and `Vdist.1compartment`.}

\item{newdata}{A `data.table` which must contain variables named `Time`,
`Dose`, `iv`, and `Media`.}

\item{DTXSID_in}{The DSSTox Substance ID for which to evaluate}

\item{Species_in}{The species for which to evaluate}

\item{Analysis_Type_in}{The analysis type that produced the fit being
evaluated: one of 'Joint', 'Separate', or 'Pooled'}

\item{Studies.Analyzed_in}{The comma-separated string of studies included in
the fit being analyzed.}

\item{model_in}{The model to evaluate: one of 'flat', '1compartment', or
'2compartment'}
}
\value{
A `data.table`, the same number of rows as `newdata`, with variables
  `pred_conc` and `pred_auc` corresponding to the model-predicted
  concentration and AUC.
}
\description{
Given a table of fitted PK parameters and a table of input data (time, dose,
route, and medium), get predicted concentrations and AUCs.
}
\author{
Caroline Ring
}
