% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starts.R
\name{get_starts}
\alias{get_starts}
\title{Get starting values for fitting model parameters}
\usage{
get_starts(
  par_DF = NULL,
  model,
  fitdata,
  starts_default = data.frame(param_name = c("A", "kelim", "Vdist", "kgutabs", "Fgutabs",
    "V1", "Ralphatokelim", "Fbetaofalpha"), start_value = c(1, 0.25, 5.56, 2.19, 1, 1, 2,
    0.25), start_value_msg = "Default"),
  suppress.messages = FALSE
)
}
\arguments{
\item{par_DF}{Optional: A data.frame as produced by [get_opt_params], with a
character variable [param_name] containing parameter names, and a logical
variable [optimize_param] containing TRUE if parameter is to be fitted, and
FALSE if parameter is to be held constant. Any other variables will be
ignored. Default NULL, in which case it will be determined by calling
[get_opt_params].}

\item{model}{The name of the model whose parameters are to be estimated.}

\item{fitdata}{A data.frame: the concentration-time-dose data to be used for
fitting.}

\item{starts_default}{A data.frame with variables `param_name` and
`start_value`, defining all possible default parameter starting values
across all models. See Details.}
}
\value{
The data.frame `par_DF` with added variables `start_value`,
  containing each parameter's starting value, and `start_value_msg`,
  containing a brief message about how each starting value was calculated.
}
\description{
For a set of model parameters, get the starting guess for the optimizer (if
parameter is to be fitted) or the value at which to hold the parameter
constant (if parameter is not to be fitted).
}
\details{
Model parameters are first assigned default starting values according to
`starts_default`.

Then, the function checks to see if the package [httk] includes the necessary
TK data to parameterize a 1-compartment model for the chemical (determined by
`unique(fitdata$DTXSID)`. If so, then it assigns [httk]'s estimates for the
relevant model parameters.
}
\author{
Caroline Ring, John Wambaugh, Mitchell Teague
}
