% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upper_bounds.R
\name{get_upper_bounds}
\alias{get_upper_bounds}
\title{Get upper bounds for estimating model parameters}
\usage{
get_upper_bounds(
  fitdata,
  par_DF = NULL,
  model,
  pool_sigma = FALSE,
  upper_default = data.frame(param_name = c("A", "kelim", "Vdist", "kgutabs", "Fgutabs",
    "V1", "k12", "k21", "Fgutabs_Vdist", "Fgutabs_V1", "sigma"), upper_bound = c(Inf,
    Inf, Inf, Inf, 1, Inf, Inf, Inf, 10000, 10000, Inf), upper_bound_msg = "Default"),
  sigma_upper_from_data = FALSE,
  suppress.messages = FALSE
)
}
\arguments{
\item{fitdata}{A data.frame: the concentration-time-dose data to be used for
fitting.}

\item{par_DF}{Optional: A data.frame as produced by [get_opt_params()], with a
character variable `param_name` containing parameter names, and a logical
variable `optimize_param` containing TRUE if parameter is to be fitted, and
FALSE if parameter is to be held constant. Any other variables will be
ignored. Default NULL, in which case it will be determined by calling
[get_opt_params()].}

\item{model}{The name of the model whose parameters are to be estimated.
Currently only "flat", "1compartment", or "2compartment" is supported.
Ignored if `par_DF` is provided.}

\item{pool_sigma}{Logical: Whether to pool all data (estimate only one error
standard deviation) or not (estimate separate error standard deviations for
each reference). Default FALSE to estimate separate error SDs for each
reference. (If `fitdata` only includes one reference, `pool_sigma` will
have no effect.) Ignored if `par_DF` is provided.}

\item{upper_default}{A `data.frame` with three variables: `param_name`,
giving the names of parameters; `upper_bound`, giving the default
upper-bound values for each parameter; and `upper_bound_msg`, giving a
message about the default upper bound values. See Details for default
value.}

\item{sigma_upper_from_data}{Logical: if TRUE, override the value for "sigma"
in `upper_default` and instead use either 100 times the median
concentration, or 2 times the overall standard deviation of concentrations,
whichever is greater. If FALSE, use the upper bound for "sigma" in
`upper_default.` Default value is TRUE.}

\item{Vdist_upper_from_data}{Logical: if TRUE, override the value for "Vdist"
or "V1" in `upper_default`, and instead use `Vdist_factor` times the
maximum dose in the data, divided by the minimum concentration or LOQ in
the data. If FALSE, use the upper bound for "Vdist" or "V1" in
`upper_default`. Default value is TRUE.}

\item{Vdist_factor}{Numeric: If `Vdist_upper_from_data` is TRUE, then
`Vdist_factor` scales the ratio between the maximum dose and minimum
concentration to arrive at a theoretical upper bound for the volume of
distribution ("vdist" or "V1"). Default value is 1. If
`Vdist_upper_from_data` is FALSE, then `Vdist_factor` has no effect.}
}
\value{
A data.frame: `parDF` with additional variables `upper_bound`
  (numeric, containing the upper bound for each parameter) and
  `upper_bound_msg` (character, containing a brief message explaining how the
  upper-bound value was calculated).
}
\description{
For a set of model parameters, get the upper bounds for the optimizer (if
parameter is to be estimated).
}
\details{
The default `upper_default` `data.frame` is shown below in table format:

| param_name     | upper_bound | upper_bound_msg |
| ---------------| ----------- | --------------- |
| A              | Inf           | Default         |
| kelim          | Inf        | Default         |
| Vdist          | Inf        | Default         |
| kgutabs        | Inf        | Default         |
| Fgutabs        | 1         | Default         |
| V1             | Inf           | Default         |
| k12            | Inf         | Default         |
| k21            | Inf         | Default         |
| Fgutabs_Vdist  | 1e4    | Default         |
| Fgutabs_V1     | 1e4    | Default         |
| sigma          | Inf         | Default         |
}
\author{
Caroline Ring, John Wambaugh, Mitchell Teague
}
