% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invivoPKfit-package.R
\docType{package}
\name{invivoPKfit}
\alias{invivoPKfit}
\title{invivoPKfit: A package to fit model parameters to in vivo data}
\description{
The invivoPKfit package takes in vivo concentration-time data from TNO, RTI,
and NHEERL, and fits parameters of 1-compartment and 2-compartment models
for each chemical. \code{\link{fit_all}} is the main or "outermost" function;
start there.
}
\section{Main function}{
 fit_all
}

\author{
Caroline Ring
}
