% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{is.pk}
\alias{is.pk}
\title{Check whether an object is of class `pk`}
\usage{
is.pk(obj)
}
\arguments{
\item{obj}{The object whose class is to be tested}
}
\value{
TRUE if the object conforms to expectations for class
  `pk`, FALSE if it does not
}
\description{
Check whether an object is of class `pk`
}
\author{
Caroline Ring
}
