% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{The log-likelihood function used for fitting}
\usage{
log_likelihood(params, DT, modelfun, model)
}
\arguments{
\item{params}{A named list of parameter values}

\item{DT}{Concentration-time data for a given chemical}

\item{modelfun}{"analytic" to use the analytic model solution, "full" to use
the full ODE model}

\item{model}{The model to fit, either "1compartment" or "2compartment"
(other models not implemented for now)}
}
\value{
A log-likelihood value for the data given the parameter values in
params
}
\description{
The log-likelihood function used for fitting
}
