% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_elim_data.R
\name{make_elim_data}
\alias{make_elim_data}
\title{Get elimination-phase data}
\usage{
make_elim_data(fitdata)
}
\arguments{
\item{fitdata}{A concentration-time-dose data.frame}
}
\value{
A data.frame with the subset of `fitdata` that reflects elimination.
  Includes new variable `ValueDose`, which is `Value` divided by `Dose`.
}
\description{
A helper function used by `get_starts` to get elimination-phase data to
roughly estimate elimination rate, volume of distribution, etc. based on
linear fits to elimination-phase data
}
\details{
Removes any rows with `is.na(Value)` and/or `Dose == 0`. If IV data
available, keeps only that data. If IV data are not available but oral data are
available, keeps only the oral data.
}
