% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_model_fits.R
\name{merge_model_fits}
\alias{merge_model_fits}
\title{Merge tables from fit_all from different models and select best fitting model}
\usage{
merge_model_fits(
  fit.list,
  dtxsid.col = "",
  compound.col = "",
  cas.col = "",
  species.col = "",
  param.value.type.col = "",
  data.analyzed.col = "Data.Analyzed",
  param.value.type = "Fitted geometric mean"
)
}
\arguments{
\item{fit_list}{a named list of PK parameter tables from \code{\link{fit_all}}
-- the names should be the names of the models}
}
\value{
A data.frame with one row per chemical-species combination
}
\description{
This takes multiple tables of pharmcokinetic parameters (one table per model)
genetated by \code{\link{fit_all}}
and merges them into a table containing one row per chemical-species 
combination. Each row contains all the model
parameters estimated for that model, and a set of optimal parameters as 
identified from the AIC. This function assigns each chemical a column 
"Model" containing the name of he winning (lowest AIC) model and, from that
winning model, reports "Vdist", "kelim", "thalf", "kgutabs", "Fbioavail".
}
\author{
John Wambaugh
}
