% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncomp_avg_value.R
\name{noncomp_avg_value}
\alias{noncomp_avg_value}
\title{Helper function to take average value of replicate timepoints}
\usage{
noncomp_avg_value(data)
}
\arguments{
\item{data}{A table of concentration-time data.}
}
\value{
A data.table with average concentration values
}
\description{
Helper function to take average value of replicate timepoints
}
\author{
Christopher Cook
}
