% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.pk_optimx_settings}
\alias{pk_add.pk_optimx_settings}
\title{Add a `pk_optimx_settings` object.}
\usage{
\method{pk_add}{pk_optimx_settings}(object, pk_obj, objectname)
}
\arguments{
\item{object}{The `pk_optimx_settings` object to be added.}

\item{pk_obj}{The `pk` object to which the `pk_optimx_settings` object will be added.}

\item{objectname}{The name of the `pk_optimx_settings` object.}
}
\value{
The `pk` object, modified by adding the settings.
}
\description{
Add a `pk_optimx_settings` object.
}
\author{
Caroline Ring
}
