% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkstats.R
\name{pkstats}
\alias{pkstats}
\title{Calculates Area Under the Curve (AUC) and Peak Concentration (Cmax)}
\usage{
pkstats(in.table, Fbio.table = NULL, default.to.human = T)
}
\value{
A table of AUCs, Cmax's, and bioavailabilities
}
\description{
author John Wambaugh
}
