% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conctime.R
\name{plot_conctime}
\alias{plot_conctime}
\title{plot_conctime}
\usage{
plot_conctime(
  PK.fit.table,
  data.set,
  model,
  mean.type = "Fitted geometric mean",
  fit.factors = c("Compound", "Data.Analyzed", "Dose.nominal", "Route", "LOQ"),
  plot.split.factor = "Route",
  shape.factor = "Data.Analyzed",
  color.factor = "Data.Analyzed",
  plot.httk.pred = F,
  omit.zero = T,
  plot.loq = T,
  fit.plot.points = 25
)
}
\arguments{
\item{PK.fit.table}{Table of parameter estimates from \code(\link(fit_all))}

\item{data.set}{A data.table of concentration vs. time data for a given
chemical}

\item{model}{Analytic model to evaluate. Currently only "1compartment" or
"2compartment" are implemented.}

\item{mean.type}{(Defaults to fitted geometric mean)}

\item{fit.factors}{Add a trend-line using the fit for each combination of the
factors listed in this character vector. (defaults to Compound, 
Dose.nominal, Data.Analyzed, and Route)}

\item{plot.split.factor}{For each compound, make a separate plot for each
 unique value of this factor (defaults to Route)
#param shape.factor Indicate this factor by shape in the plots (defaulta to
 Data.Analyzed)}

\item{color.factor}{Indicate this factor by color in the plots (defaults to
Data.Analyzed')}

\item{omit.zero}{Drop zero dose values from plots (default = TRUE)}

\item{Add}{a curve for the HTTK predictions, if available (default = FALSE)}
}
\value{
None
}
\description{
Creates a PDF of the fits from \code{\link{fit_all}} and data together
}
\author{
{Caroline Ring, John Wambaugh}
}
