% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot TK model fits}
\usage{
plot_fit(
  fit_flat,
  fit_1comp,
  fit_2comp,
  DTXSID_in,
  Species_in,
  Analysis_Type_in,
  Studies.Analyzed_in,
  model_in = "winning",
  obs_data,
  plot_dose_norm = TRUE,
  split_dose = FALSE,
  n_interp_time = 10,
  limit_y_axis = FALSE,
  log10_scale_y = FALSE,
  plot_errbars = TRUE,
  log10_scale_time = FALSE,
  jitter_nondetect = 0.1,
  save_plot = TRUE,
  return_plot = FALSE,
  file_path = "inst/ext/plots/",
  file_suffix = NULL,
  file_format = "pdf",
  ggsave_args = list(scale = 1, width = 9, height = 10, units = "in", dpi = 300,
    limitsize = TRUE, bg = NULL),
  verbose = TRUE
)
}
\arguments{
\item{DTXSID_in}{DSSTox Substance ID of chemical to be plotted.}

\item{Species_in}{Species in data to be plotted}

\item{Analysis_Type_in}{"Joint", "Pooled", or "Separate" (whether different
studies were fit together with separate error SDs; together with one error
SD; or separately)}

\item{model_in}{Which model or models to plot fits for. Options include any or
all of "1compartment", "2compartment", and "flat"; "winning" to plot only
the winning model (whichever had lowest AIC); or "all" as a short form to
plot all three models.}

\item{obs_data}{A `data.table`: Observed concentration vs. time data to be
plotted, pre-processed as from [preprocess_data()]. Must have variables
`DTXSID`, `Species`, `Route`, `Media`, `iv`, `Dose`, `Time`, `Conc`,
`Detect`, `Conc_SD`.}

\item{plot_dose_norm}{Logical: TRUE to plot dose-normalized concentration vs.
time data and model fits. FALSE to plot non-dose-normalized data and fits.
Default TRUE. If `plot_dose_norm == FALSE`, it's typically recommended to
use `split_dose == TRUE`, because non-dose-normalized concentrations for
different dose groups will likely be on different scales.}

\item{split_dose}{Logical: TRUE to split separate dose groups into separate
plot facets. FALSE to facet only by Route and Media. If `split_dose ==
FALSE`, then dose groups will be distinguished using color.}

\item{n_interp_time}{The number of time points to interpolate between each
existing time point in the data set for plotting model fits. Default is 10.}

\item{limit_y_axis}{Whether to limit the scale of the y axis to only a small
distance above and below the min/max of the concentration vs. time data.
Default FALSE. (You may want to try TRUE if you are log-scaling the y axis.)}

\item{log10_scale_y}{Logical: Whether to log-scale the y axis. Default FALSE.}

\item{plot_errbars}{Logical: Whether to plot sample errorbars for
concentration vs. time data (when the observations have associated sample
standard deviations). Default TRUE. You may want to try FALSE if you are
log-scaling the y axis.}

\item{log10_scale_time}{Logical: Whether to log-scale the x-axis (time).
Default FALSE.}

\item{save_plot}{Logical: Whether to save the plot as a file. Default TRUE.}

\item{return_plot}{Logical: Whether to return the plot object. Default FALSE.}

\item{file_path}{Path to save file if `save_plot == TRUE`. Default
`"inst/ext/plots"`.}

\item{file_suffix}{Suffix to append to the file name, if any. Default NULL for
no suffix.}

\item{file_format}{File format to save the plot. Default `"pdf"`.}

\item{ggsave_args}{A named list of additional arguments to `ggsave()` to use
 when saving the plot. Default
  ```
  list(scale = 1,
width = 9,
height = 10,
units = "in",
dpi = 300,
limitsize = TRUE,
bg = NULL)
```}

\item{verbose}{Logical: Whether to print verbose messages about plotting
progress. Default TRUE.}
}
\value{
If `return_plot == TRUE`, a `ggplot2` object. If `return_plot ==
 FALSE`, returns 0.
}
\description{
Plot fitted toxicokinetic model curves with concentration vs. time data
}
\details{
This function can plot dose-normalized data, non-dose-normalized data, can
split doses into separate facets or not.
}
\author{
Caroline Ring
}
