% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resids.R
\name{plot_resids}
\alias{plot_resids}
\title{Plot residuals from winning model}
\usage{
plot_resids(
  cvt_pre,
  fit_flat,
  fit_1comp,
  fit_2comp,
  return_plots = FALSE,
  save_indiv_plots = TRUE,
  save_combined_plots = TRUE,
  file_path = "../inst/ext/plots/resid_plots",
  file_suffix = NULL,
  file_format = "pdf",
  ggsave_args = list(scale = 1, width = 13.33, height = 7.5, units = "in", dpi = 300,
    limitsize = TRUE, bg = NULL)
)
}
\arguments{
\item{cvt_pre}{Original concentration vs. time data, pre-processed using
`preprocess_data()`.}

\item{fit_flat}{Fitting output for the flat model: the output of `fit_all()`
with `model = "flat"`}

\item{fit_1comp}{Fitting output for the 1-compartment model: the output of
`fit_all()` with `model = "1compartment"`}

\item{fit_2comp}{Fitting output for the 2-compartment model: the output of
`fit_all()` with `model = "2compartment"`}
}
\value{
A `data.table` of goodness-of-fit measures for each dataset, each
  model, and each analysis (joint, separate, and pooled)
}
\description{
Make diagnostic plots of residuals from the winning TK model for a dataset
}
\details{
For a given dataset (DTXSID and Species), this function calculates the
predicted concentrations from the winning (best-fit by lowest AIC) TK model.
Then it computes the residuals (observed - predicted), both dose-normalized
and non-dose-normalized. Then it creates a grid of plots with two rows and
three columns. The first row shows non-dose-normalized residuals; the second
row shows dose-normalized residuals. The first column plots residuals vs.
observed concentration. The second column plots residuals vs. dose. The third
column plots residuals vs. time.

These plots can be used to diagnose heteroscedasticity in residual errors
from the best-fit model.
}
\author{
Caroline Ring
}
