% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_data.R
\name{postprocess_data}
\alias{postprocess_data}
\title{Postprocess fitted PK models}
\usage{
postprocess_data(PK_fit, model)
}
\arguments{
\item{PK_fit}{A data.table of fitted PK parameters, as produced by `fit_all()` or `analyze_subset()`}

\item{model}{The name of the relevant model: "1compartment", "2compartment", or "flat"}
}
\value{
A data.table of derived quantities relevant to the model, with one row for each set of model parameters
}
\description{
Derive quantities such as half-life, tmax, Cmax, Css from fitted PK models
}
\author{
John Wambaugh, Caroline Ring
}
