% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_conc.R
\name{predict_conc}
\alias{predict_conc}
\title{Get model predicted concentration and AUC}
\usage{
predict_conc(pk_fit_row, newdata, model)
}
\arguments{
\item{newdata}{A `data.table` which must contain variables named `Time`,
`Dose`, `iv`, and `Media`.}

\item{model}{The model to evaluate: one of 'flat', '1compartment', or
'2compartment'}

\item{pk_fit}{A table of fitted PK parameters, as produced by
`postprocess_data()`. Must contain variables `DTXSID`, `Species`,
`Analysis_Type`, and `Studies.Analyzed`. Must also contain variables
corresponding to the parameters of the model specified in argument
`model_in`, as required for the route and media of data in `newdata` (as
given by [get_model_paramnames()]). These variables must be named as
`[param].[model]`. For example, for the 1-compartment model, if `newdata`
contains only IV-dosing data measured in plasma, then `pk_fit` must contain
variables named `kelim.1compartment` and `Vdist.1compartment`.}
}
\value{
A `data.table`, the same number of rows as `newdata`, with variables
  `pred_conc` and `pred_auc` corresponding to the model-predicted
  concentration and AUC.
}
\description{
Given a table of fitted PK parameters and a table of input data (time, dose,
route, and medium), get predicted concentrations and AUCs.
}
\author{
Caroline Ring
}
