% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{prefit.pk}
\alias{prefit.pk}
\title{Do pre-fit calculations and checks}
\usage{
prefit.pk(obj)
}
\arguments{
\item{obj}{A `pk` object}
}
\value{
The same `pk` object, but with additional elements added to each item
  in `models`, containing the results of pre-fit calculations and checks for
  each model.
}
\description{
Do pre-fit calculations and checks
}
