% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{preprocess_data.pk}
\alias{preprocess_data.pk}
\title{Pre-process data}
\usage{
\method{preprocess_data}{pk}(obj)
}
\arguments{
\item{obj}{A `pk` object}
}
\value{
The same `pk` object, with added elements `data` (containing the
  cleaned, gap-filled data) and `data_info` (containing summary information
  about the data, e.g. number of observations by route, media,
  detect/nondetect; empirical tmax, time of peak concentration for oral data;
  number of observations before and after empirical tmax)
}
\description{
S3 method for `pk` objects to preprocess data
}
\details{
This is the `preprocess_data` S3 method for objects of class `pk`.
}
\author{
Caroline Ring
}
