% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{rescale_time.pk}
\alias{rescale_time.pk}
\title{Rescale time}
\usage{
rescale_time.pk(obj)
}
\arguments{
\item{obj}{A [pk()] object.}
}
\value{
The `data` element of the object, but modified: the original
  time (in hours) is now in variable `Time.Hours`; rescaled time is in
  `Time`; units of rescaled time are in `Time.Units`.
}
\description{
If requested, rescale time from hours to some more convenient unit.
}
\details{
Time is rescaled based upon the time of the last detected observation in units of hours.

 | Range of Last Detect Time in Hours | New Time Units | Range of New Last Detect Time |
 | ----------- | -------------- |
 | \\[0, 0.5) | minutes | \\[0, 30) |
 | \\[0.5, 24) | hours | \\[0.5, 24) |
 | \\[24, 720) | days | \\[1, 30) |
 | \\[720, 8766) | months | \\[0.99, 12) |
 | \\[8766, 87660) | years | \\[1, 10) |
 | \\[87660, 876600) | decades | \\[1, 10) |
 | \\[876600, Inf) | centuries | \\[1, Inf) |

 (At present, no data in CvT requires decades or centuries)
}
\author{
Caroline Ring
}
