% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{residuals.pk}
\alias{residuals.pk}
\title{Get residuals}
\usage{
\method{residuals}{pk}(obj, newdata = NULL, model = NULL, method = NULL, type = "conc")
}
\arguments{
\item{obj}{A `pk` object}

\item{newdata}{Optional: A `data.frame` with new data for which to make
predictions and compute residuals. If NULL (the default), then residuals
will be computed for the data in `obj$data`. `newdata` is required to
contain at least the following variables: `Time`, `Dose`, `Route`, and
`Media`. `Time` will be transformed according to the transformation in
`obj$scales$time` before residuals are calculated.}

\item{model}{Optional: Specify one or more of the fitted models for which to
make predictions and calculate residuals. If NULL (the default), residuals
will be returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to make predictions and calculate residuals. If NULL (the
default), residuals will be returned for all of the models in
`obj$optimx_settings$method`.}

\item{type}{Either `"conc"` (the default) or `"auc"`. `type = "conc"`
predicts concentrations; `type = "auc"` predicts area under the
concentration-time curve (AUC). Currently, only `type = "conc"` is
implemented.}
}
\value{
A named list of numeric matrices. There is one list element named for
  each model in `obj`'s [stat_model()] element, i.e. each PK model that was
  fitted to the data. Each list element is a matrix with the same number of
  rows as the data in `obj$data` (corresponding to the rows in `obj$data`),
  and as many columns as there were [optimx::optimx()] methods (specified in
  [settings_optimx()]). The column names are the method names.  Each column
  contains the residuals (observed - predicted) of the model fitted by the
  corresponding method. These residuals are concentrations in the same units
  as `obj$data$Conc.Units`; any concentration transformations (in
  `obj$scale$conc`) are *not* applied.
}
\description{
Extract residuals from a fitted `pk` object.
}
\details{
Residuals are `observed - predicted`.
}
\author{
Caroline Ring
}
