% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{rmse.pk}
\alias{rmse.pk}
\title{Root mean squared error}
\usage{
\method{rmse}{pk}(obj, newdata = NULL, model = NULL, method = NULL, type = "conc")
}
\arguments{
\item{obj}{A `pk` object}

\item{newdata}{Optional: A `data.frame` with new data for which to make
predictions and compute RMSsE. If NULL (the default), then RMSEs will be
computed for the data in `obj$data`. `newdata` is required to contain at
least the following variables: `Time`, `Dose`, `Route`, and `Media`. `Time`
will be transformed according to the transformation in `obj$scales$time`
before RMSEs are calculated.}

\item{model}{Optional: Specify one or more of the fitted models for which to
make predictions and calculate RMSEs. If NULL (the default), RMSEs will be
returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to make predictions and calculate RMSEs. If NULL (the default),
RMSEs will be returned for all of the models in
`obj$optimx_settings$method`.}

\item{type}{Either `"conc"` (the default) or `"auc"`. `type = "conc"`
predicts concentrations; `type = "auc"` predicts area under the
concentration-time curve (AUC). Currently, only `type = "conc"` is
implemented.}
}
\value{
A named list of numeric vectors. There is one list element named for
  each model in `obj`'s [stat_model()] element, i.e. each PK model that was
  fitted to the data. Each list element is a numeric vector with as many
  elements as there were [optimx::optimx()] methods (specified in
  [settings_optimx()]). The vector names are the method names.  Each vector
  element contains the root mean squared error (`sqrt(mean(observed -
  predicted)^2)`) of the model fitted by the corresponding method, using the
  data in `newdata`. These RMSEs are concentrations in the same units as
  `obj$data$Conc.Units`; any concentration transformations (in
  `obj$scale$conc`) are *not* applied.
}
\description{
Extract root mean squared error of a fitted `pk` object
}
\author{
Caroline Ring
}
