% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_data_settings.R
\name{settings_optimx}
\alias{settings_optimx}
\title{`optimx` optimizer settings}
\usage{
settings_optimx(
  method = "bobyqa",
  itnmax = 1e+06,
  hessian = FALSE,
  control = list(kkt = FALSE),
  ...
)
}
\arguments{
\item{method}{The name(s) of optimization methods to be used. See
[optimx::optimx()] for options. Default is `"bobyqa"`.}

\item{itnmax}{The maximum number of iterations; as in [optimx::optimx()].}

\item{hessian}{Whether to compute the Hessian at the final set of parameters; as in [optimx::optimx()].}

\item{control}{A list of control parameters for the optimizer; see [optimx::optimx()] for options and details.}
}
\value{
An object of class `pk_settings`.
}
\description{
`optimx` optimizer settings
}
\author{
Caroline Ring
}
