% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_stat_model.R
\name{stat_error_model}
\alias{stat_error_model}
\title{Error model}
\usage{
stat_error_model(error_group = vars(Chemical, Species, Reference, Media), ...)
}
\arguments{
\item{error_group}{Defined using [ggplot2::vars()]: A set of harmonized
variables whose unique combinations define a group with its own error
variance. These variables refer to the `data` element of the `pk` object to
which `stat_error_model()` is added.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class `pk_stat_error_model`: A named list of all the
 arguments to `stat_error_model`.
}
\description{
Define an error model.
}
\author{
Caroline Ring
}
