% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_1comp.R
\name{cp_1comp}
\alias{cp_1comp}
\title{Analytical 1-compartment model}
\usage{
cp_1comp(params, time, dose, iv.dose)
}
\arguments{
\item{params}{A named list of model parameter values. For either IV or oral
dosing, this must include `kelim` (elimination rate, 1/h). For IV dosing, it
must also include `Vdist`. For oral dosing, it must include `kgutabs` (oral
absorption rate, 1/h), and either `Fgutabs` and `Vdist` (respectively,
unitless fraction of oral dose that is absorbed, and volume of
distribution), or `Fgutabs_Vdist` (the ratio of Fgutabs to Vdist).
`Fgutabs_Vdist` is an alternate parameterization useful for parameter
estimation when only oral data are available with no IV data, in which case
only the ratio of `Fgutabs` to `Vdist` is identifiable. if `Fgutabs` and
`Vdist` are provided along with `Fgutabs_Vdist`, then `Fgutabs_Vdist` will
not be used.}

\item{time}{A vector of times in hours.}

\item{dose}{Dose in mg/kg}

\item{iv.dose}{Logical: TRUE for single IV bolus dose; FALSE for single oral
dose}
}
\value{
A vector of plasma concentration values corresponding to `time`.
}
\description{
Calculates plasma concentrations according to the analytical solution for the
1-compartment model.
}
\author{
Caroline Ring, John Wambaugh
}
