% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_1comp.R
\name{cp_1comp}
\alias{cp_1comp}
\title{Analytical 1-compartment model}
\usage{
cp_1comp(time, params, dose, iv.dose)
}
\arguments{
\item{time}{A vector of times in hours}

\item{params}{A named list of model parameter values. Must include:
\describe{
\item{kelim}{Elimination rate, 1/h}
\item{Vdist}{Volume of distribution, L/kg body weight}}
For oral administration (\code{iv.dose} FALSE), \code{params} must also include:
\describe{
\item{Fgutabs}{Oral bioavailability, unitless fraction}
\item{kgutabs}{Oral absorption rate, 1/h}}}

\item{dose}{Dose in mg/kg}

\item{iv.dose}{TRUE for single IV bolus dose; FALSE for single oral dose}
}
\value{
A vector of plasma concentration values corresponding to \code{time}.
}
\description{
Analytical 1-compartment model
}

