% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_flat.R
\name{cp_flat}
\alias{cp_flat}
\title{Flat model}
\usage{
cp_flat(time, params, dose, iv.dose)
}
\arguments{
\item{time}{A numeric vector of times in hours.}

\item{params}{A named list of model parameter values. Must include: \describe{
\item{A}{Average concentration} }}

\item{dose}{A numeric vector of doses in mg/kg}

\item{iv.dose}{A logical vector: TRUE for single IV bolus dose; FALSE for
single oral dose. Not used, but must be present for compatibility with other
model functions.}
}
\value{
A vector of plasma concentration values (mg/L) corresponding to \code{time}.
}
\description{
Flat model
}
\author{
Caroline Ring, John Wambaugh, Chris Cook
}
