% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_loq.R
\name{estimate_loq}
\alias{estimate_loq}
\title{Assign an estimated limit of quantification wherever missing}
\usage{
estimate_loq(
  dat,
  reference_col = "Reference",
  chem_col = "DTXSID",
  media_col = "Media",
  species_col = "Species",
  value_col = "Value",
  loq_col = "LOQ",
  calc_loq_factor = 0.45
)
}
\arguments{
\item{dat}{A \code{data.frame} of concentration vs. time data with at least
the following variables:}

\item{reference_col}{The name of the variable in `dat` indicating different
research study reference documents (and therefore different mass
spectrometry methods). Default "Reference".}

\item{chem_col}{The name of the variable in `dat` indicating chemical
identity. Default "DTXSID".}

\item{media_col}{The name of the variable in `dat` indicating sample medium.
Default "Medium".}

\item{species_col}{The name of the variable in `dat` indicating species.
Default "Species".}

\item{value_col}{The name of the variable in `dat` indicating concentration.
Default "Value". Nondetects should be reported as NA in this column.}

\item{loq_col}{The name of the LOQ column. Default "LOQ". If this column does
not already exist, it will be created and filled with `NA_real_`.}

\item{calc_loq_factor}{A factor by which to multiply the lowest detected
value in each group to estimate an LOQ. Default 0.45.}
}
\value{
A \code{data.frame} that is the same as \code{dat}, but with any
  missing values in variable `calc_loq` imputed as above.
}
\description{
Sometimes in mass spectrometry a compound is detected, but the signal is too
low and falls outside the range that can be interpreted quantitatively (that
is, as a concentration). These detects can still be used to refine parameter
estimates by preferring curve fits that predict concentrations below the
known "limit of quantitation" (LOQ) -- the lowest detectable quantifiable
concentration. LOQs are not always available from data reported in the
literature. When LOQ is missing, this function imputes an LOQ such that all
the reported values are above the LOQ.
The LOQ is assumed to vary from reference to reference, as each study would
have different mass spectrometry equipment and methods. LOQ also is assumed
to vary by medium (e.g. blood or plasma), as preparation of samples may be
different for different media.
}
\details{
For each unique combination of reference, chemical, species, and medium, the
LOQ is estimated by multiplying the lowest detected concentration (the
minimum non-NA reported concentration greater than zero) by a constant factor
(`calc_loq_factor`). If there are no detected concentrations for a given
combination of reference, chemical, species, and medium (*i.e.*, all reported
concentrations are NA), then the imputed LOQ will be NA.
}
\author{
John Wambaugh, Caroline Ring
}
