% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_resids.R
\name{evaluate_resids}
\alias{evaluate_resids}
\title{Evaluate goodness of fit based on residuals}
\usage{
evaluate_resids(
  cvt_pre,
  pk_fit_row,
  optimx_args = list(method = "bobyqa", itnmax = 1e+06, control = list(kkt = FALSE)),
  verbose = TRUE
)
}
\arguments{
\item{cvt_pre}{Concentration vs. time data for which to evaluate residuals.
Should be a subset of the table produced by [preprocess_data()], containing
data for the DTXSID and species corresponding to `pk_fit_row`.}

\item{pk_fit_row}{One row of post-processed fitting output, as produced by
[merge_fits()]. This corresponds to one DTXSID, species, analysis type, and
one set of fitting options.}
}
\value{
A `data.table` of goodness-of-fit measures for each dataset, each
  model, and each analysis (joint, separate, and pooled)
}
\description{
Evaluate multiple metrics of goodness of fit based on residuals
}
\details{
For each dataset (chemical and species) and each fitted model, this function
first calculates the predictions and residuals, then evaluates the following:

- RMSE (root mean squared error)
- R-squared of observations vs. predictions
- Estimated "overall" error standard deviation (details below)

Unique fitted models are defined by unique combination of analysis type
("Joint", "Separate", or "Pooled"), studies analyzed, and model ("flat",
"1compartment", or "2compartment").

# Root mean squared error (RMSE)

RMSE is calculated in [calc_rmse()].

# Estimated overall error standard deviation

The "overall" error standard deviation is the value for residual error
standard deviation ("sigma") that minimizes the log-likelihood function when
the TK model parameters are fixed at their fitted values. This is essentially
a more rigorous RMSE, because it correctly models the likelihood of
non-detect (below-LOQ) observations as the cumulative probability that the
model prediction is below the reported LOQ, rather than making any
substitutions for non-detects.
}
\author{
Caroline Ring
}
