% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all.R
\name{fit_all}
\alias{fit_all}
\title{Main fitting function}
\usage{
fit_all(
  data.set,
  model,
  modelfun = NA,
  names_list = list(Compound_Dosed = "studies.test_substance_name_original", DTXSID_Dosed
    = "chemicals_dosed.dsstox_substance_id", CAS_Dosed = "chemicals_dosed.dsstox_casrn",
    Compound_Analyzed = "series.analyte_name_original", DTXSID_Analyzed =
    "chemicals_analyzed.dsstox_substance_id", CAS_Analyzed =
    "chemicals_analyzed.dsstox_casrn", Reference = "documents_reference.id", Extraction =
    "documents_extraction.id", Species = "subjects.species", Weight =
    "subjects.weight_kg", Weight.Units = NULL, Dose = "studies.dose_level_normalized", 
 
       Dose.Units = NULL, Time = "conc_time_values.time_hr", Time.Units = NULL, Media =
    "series.conc_medium_normalized", Value = "conc_time_values.conc", Value.Units = NULL,
    Route = "studies.administration_route_normalized", LOQ = "series.loq", Subject =
    "subjects.id", N_Subjects = "series.n_subjects_in_series", Value_SD =
    "conc_time_values.conc_sd"),
  defaults_list = list(Weight.Units = "kg", Dose.Units = "mg/kg", Time.Units = "h",
    Value.Units = "mg/L"),
  ratio_conc_to_dose = 1,
  calc_loq_factor = 0.45,
  routes_keep = c("po", "iv"),
  media_keep = c("blood", "plasma"),
  get_starts_args = NULL,
  get_lower_args = NULL,
  get_upper_args = NULL,
  optimx_args = list(method = "bobyqa", itnmax = 1e+06, control = list(kkt = FALSE)),
  suppress.messages = FALSE
)
}
\arguments{
\item{data.set}{A \code{data.frame} of concentration-time data. Preferably
\code{pkdataset_nheerlcleaned} or \code{pkdataset_nheerlorig}.}

\item{model}{Which general model should be fit for each chemical. Presently,
only "1compartment" and "2compartment" are implemented.}

\item{modelfun}{Either "analytic" or "full" -- whether to fit using the
analytic solution to the model, or the full ODE model. Presently,
"analytic" is recommended (because the analytic solution is exact and much
faster).}

\item{names_list}{As for [rename_columns()]: A named list where the names are
the new variable names, and the values are the old variable names. If a
value is NULL, then there is no old variable corresponding to that new
variable. A new variable will be added, with default value as given in
`defaults_list`; if no default is given in `defaults_list`, the new
variable will be filled with `NA_character`.}

\item{defaults_list}{As for [rename_columns()]: A named list where the names
are the new variable names, and the values are the default values to fill
for each new variable, if the corresponding old variable name is NULL in
`names_list`.}

\item{ratio_conc_to_dose}{Ratio between the mass units used to report the
concentration data and the mass units used to report the dose. Default 1.
For example, concentration reported in ug/L and dose reported in mg/kg/day
would require `ratio_conc_to_dose = 0.001`, because 1 ug/1 mg = 1e-6 g /
1e-3 g = 0.001.}

\item{calc_loq_factor}{As for [estimate_loq()]}

\item{routes_keep}{Character vector of administration routes to keep. Default
`c("po", "iv")` to keep only oral and intravenous data (because at present,
model fitting is only implemented for oral and IV routes). Any data where
route is not on this list will be discarded. (For example, with the
default, inhalation and dermal data will be discarded.)}

\item{media_keep}{Character vector of sample media to keep. Default is
`c("blood", "plasma")` to keep only concentration data measured in blood
and plasma (because at present, model fitting is only implemented for blood
and plasma). Any data where medium is not on this list will be discarded.
(For example, with the default, urine data will be discarded.)}

\item{get_starts_args}{Named list or NULL: any additional arguments to
[get_starts()] (other than `model` and `fitdata`, which are always passed).
Default NULL to accept the default arguments for [get_starts()].}

\item{get_lower_args}{Named list or NULL: any additional arguments to
[get_lower_bounds()] (other than `model` and `fitdata`, which are always
passed). Default NULL to accept the default arguments for
[get_lower_bounds()].}

\item{get_upper_args}{Named list or NULL: any additional arguments to
[get_upper_bounds()] (other than `model` and `fitdata`, which are always
passed). Default NULL to accept the default arguments for
[get_upper_bounds()].}

\item{optimx_args}{A named list of additional arguments to [optimx::optimx()]
  (used for parameter estimation), other than `par`, `fn`, `lower`, and
  `upper`. Default is:

   ```
    list(
          "method" = "bobyqa",
          "itnmax" = 1e6,
          "control" = list("kkt" = FALSE)
         )
   ```
  See documentation for [optimx::optimx()] for possible arguments and
  details. Note lower and upper bounds (box constraints) will be supplied; if
  you want them to be respected, please choose a method that allows box
  constraints (e.g. "bobyqa" or "L-BFGS-B").}

\item{suppress.messages}{Logical: Whether to suppress verbose messages.
Default FALSE, to be verbose.}
}
\value{
A `data.table` of fitted parameter values for each chemical.
}
\description{
Fits parameters of a specified model to concentration-time data.
}
\author{
Caroline Ring, John Wambaugh
}
