% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all.R
\name{fit_all}
\alias{fit_all}
\title{Main fitting function}
\usage{
fit_all(
  data.set,
  model,
  modelfun = NA,
  ratio.data.to.dose = 1,
  compound.col = "Compound",
  dtxsid.col = "DTXSID",
  cas.col = "CAS",
  reference.col = "Reference",
  species.col = "Species",
  species.default = NULL,
  species.weight.col = "Species.Weight",
  species.weight.units.col = "Species.Weight.Units",
  species.weight.units.default = NULL,
  dose.col = "Dose",
  time.col = "Time",
  time.units.col = "Time.Units",
  time.units.default = NULL,
  media.col = "Media",
  media.units.col = "Media.Units",
  media.units.default = NULL,
  value.col = "Value",
  units.col = "Units",
  units.default = NULL,
  route.col = "Route",
  route.default = NULL,
  source.col = "Source",
  source.default = NULL,
  loq.col = "LOQ",
  loq.default = NULL,
  subject.col = "Subject",
  subject.default = NULL,
  info.col = "info",
  info.default = NULL,
  LOQ_factor = 2,
  get_starts_args = NULL,
  get_lower_args = NULL,
  get_upper_args = NULL,
  optimx_args = list(method = "bobyqa", itnmax = 1e+06, control = list(maximize = TRUE,
    kkt = FALSE)),
  suppress.messages = FALSE
)
}
\arguments{
\item{data.set}{A \code{data.frame} of concentration-time data. Preferably
\code{pkdataset_nheerlcleaned} or \code{pkdataset_nheerlorig}.}

\item{model}{Which general model should be fit for each chemical. Presently,
only "1compartment" and "2compartment" are implemented.}

\item{modelfun}{Either "analytic" or "full" -- whether to fit using the
analytic solution to the model, or the full ODE model. Presently,
"analytic" is recommended (because the analytic solution is exact and much
faster).}

\item{ratio.data.to.dose}{Ratio between the mass units used to report the
concentration data and the mass units used to report the dose. Default 1.
For example, concentration reported in ug/L and dose reported in mg/kg/day
would require \code{ratio.data.to.dose = 0.001}, because 1 ug/1 mg = 1e-6 g
/ 1e-3 g = 0.001.}

\item{compound.col}{Column name in \code{data.set} that identifies chemical
compound. Default "Compound".}

\item{cas.col}{Column name in \code{data.set} to identify CASRN.
Default"CAS".}

\item{reference.col}{Column name in \code{data.set} to identify reference.
Default "Reference".}

\item{species.col}{Column name in \code{data.set} to identify species.
Default "Species".}

\item{species.default}{If no species column exists in \code{data.set}, one
will be created and filled with this value.  Default NULL.}

\item{species.weight.col}{Column name in \code{data.set} to identify species
weight. Default "Species.Weight".}

\item{species.weight.units.col}{Column name in \code{data.set} to identify
species weigh units. Default"Species.Weight.Units".}

\item{species.weight.units.default}{If no species weight units column exists
in \code{data.set}, one will be created and filled with this value. Default
NULL.}

\item{dose.col}{Column name in \code{data.set} to identify dose. Default
"Dose".}

\item{time.col}{Column name in \code{data.set} to identify time. Default
"Time".}

\item{time.units.col}{Column name in \code{data.set} to identify time units.
Default "Time.Units."}

\item{time.units.default}{If no time units column exists in \code{data.set},
one will be created and filled with this value.  Default NULL.}

\item{media.col}{Column name in \code{data.set} to identify media. Default
"Media".}

\item{media.units.col}{Column name in \code{data.set} to identify media
units. Default "Media.Units".}

\item{media.units.default}{If no media units column exists in
\code{data.set}, one will be created and filled with this value.  Default
NULL.}

\item{value.col}{Column name in \code{data.set} to identify value. Default
"Value".}

\item{units.col}{Column name in \code{data.set} to identify units. Default
"Units".}

\item{units.default}{If no units column exists in \code{data.set}, one will
be created and filled with this value.  Default NULL.}

\item{route.col}{Column name in \code{data.set} to identify route of
administration. Default "Route".}

\item{route.default}{If no route column exists in \code{data.set}, one will
be created and filled with this value.  Default NULL.}

\item{source.col}{Column name in \code{data.set} to identify source. Default
"Source."}

\item{source.default}{If no source column exists in \code{data.set}, one will
be created and filled with this value.  Default NULL.}

\item{loq.col}{Column name in \code{data.set} to identify LOQ. Default "LOQ".}

\item{loq.default}{If no LOQ column exists in \code{data.set}, one will be
created and filled with this value.  Default NULL.}

\item{subject.col}{Column name in \code{data.set} to identify subject.
Default "Subject."}

\item{subject.default}{If no subject column exists in \code{data.set}, one
will be created and filled with this value.  Default NULL.}

\item{info.col}{Column name in \code{data.set} to serve as info column.
Default "Info".}

\item{info.default}{If no info column exists in \code{data.set}, one will be
created and filled with this value.  Default NULL.}

\item{LOQ_factor}{Numeric. Observations with concentrations less than
`LOQ_factor * LOQ` will be removed. Default 2.}

\item{get_starts_args}{Named list or NULL: any additional arguments to [get_starts()] (other than
`model` and `fitdata`, which are always passed). Default NULL to accept the
default arguments for [get_starts()].}

\item{get_lower_args}{Named list or NULL: any additional arguments to [get_lower_bounds()] (other
than `model` and `fitdata`, which are always passed). Default NULL to
accept the default arguments for [get_lower_bounds()].}

\item{get_upper_args}{Named list or NULL: any additional arguments to [get_upper_bounds()] (other
than `model` and `fitdata`, which are always passed). Default NULL to
accept the default arguments for [get_upper_bounds()].}

\item{optimx_args}{A named list of additional arguments to [optimx::optimx()]
  (used for parameter estimation), other than `par`, `fn`, `lower`, and
  `upper`. Default is:

   ```
    list(
          "method" = "bobyqa",
          "itnmax" = 1e6,
          "control" = list("maximize" = TRUE,
                           "kkt" = FALSE)
         )
   ```
 See documentation for [optimx::optimx()] for possible arguments and details.
 Note lower and upper bounds (box constraints) will be supplied; if you want
 them to be respected, please choose a method that allows box constraints
 (e.g. "bobyqa" or "L-BFGS-B").}

\item{suppress.messages}{Logical: Whether to suppress verbose messages.
Default FALSE, to be verbose.}
}
\value{
A `data.table` of fitted parameter values for each chemical.
}
\description{
Fits parameters of a specified model to concentration-time data.
}
\author{
Caroline Ring, John Wambaugh
}
