% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all.R
\name{fit_all}
\alias{fit_all}
\title{Main fitting function}
\usage{
fit_all(
  data.set,
  model,
  modelfun = NA,
  ratio.data.to.dose = 1,
  compound.col = "Compound",
  cas.col = "CAS",
  reference.col = "Reference",
  species.col = "Species",
  species.default = NULL,
  species.weight.col = "Species.Weight",
  species.weight.units.col = "Species.Weight.Units",
  species.weight.units.default = NULL,
  dose.col = "Dose",
  time.col = "Time",
  time.units.col = "Time.Units",
  time.units.default = NULL,
  media.col = "Media",
  media.units.col = "Media.Units",
  media.units.default = NULL,
  value.col = "Value",
  units.col = "Units",
  units.default = NULL,
  route.col = "Route",
  route.default = NULL,
  source.col = "Source",
  source.default = NULL,
  loq.col = "LOQ",
  loq.default = NULL,
  subject.col = "Subject",
  subject.default = NULL,
  info.col = "info",
  info.default = NULL
)
}
\arguments{
\item{data.set}{A table of concentration-time data. Preferably
\code{pkdataset_nheerlcleaned} or \code{pkdataset_nheerlorig}.}

\item{model}{Which general model should be fit for each chemical. Presently,
only "1compartment" and "2compartment" are implemented.}

\item{modelfun}{Either "analytic" or "full" -- whether to fit using the
analytic solution to the model, or the full ODE model. Presently,
"analytic" is recommended (because the analytic solution is exact and much
faster).}

\item{ratio.data.to.dose}{between the weights used to report the data and the weights used
for the dose. For example, ug/L data and mg/kg/day dose would be 0.001
(defaults to 1)}

\item{compound.col}{Column in data.set to serve as compound column}

\item{cas.col}{Column in data.set to serve as CAS column}

\item{reference.col}{Column in data.set to serve as reference column}

\item{species.col}{Column in data.set to serve as species column}

\item{species.default}{Default value to fill newly created species column, should data.set lack a column equivalent to species}

\item{species.weight.col}{Column in data.set to serve as species.weight column}

\item{species.weight.units.col}{Column in data.set to serve as species.weight.units column}

\item{species.weight.units.default}{Default value to fill newly created species.weight.units column, should data.set lack a column equivalent to species.weight.units}

\item{dose.col}{Column in data.set to serve as dose column}

\item{time.col}{Column in data.set to serve as time column}

\item{time.units.col}{Column in data.set to serve as time.units column}

\item{time.units.default}{Default value to fill newly created time.units column, should data.set lack a column equivalent to time.units}

\item{media.col}{Column in data.set to serve as media column}

\item{media.units.col}{Column in data.set to serve as media.units column}

\item{media.units.default}{Default value to fill newly created media.units coloumn, should data.set lack a column equivalent to media.units}

\item{value.col}{Column in data.set to serve as value column}

\item{units.col}{Column in data.set to serve as units column}

\item{units.default}{Default value to fill newly created units column, should data.set lack a column equivalent to units}

\item{route.col}{Column in data.set to serve as route column}

\item{route.default}{Default value to fill newly created route column, should data.set lack a column equivalent to route}

\item{source.col}{Column in data.set to serve as source column}

\item{source.default}{Default value to fill newly created source column, should data.set lack a column equivalent to source}

\item{loq.col}{Column in data.set to serve as loq column}

\item{loq.default}{Default value to fill newly created loq column, should data.set lack a column equivalent to loq}

\item{subject.col}{Column in data.set to serve as subject column}

\item{subject.default}{Default value to fill newly created subject column, should data.set lack a column equivalent to subject}

\item{info.col}{Column in data.set to serve as info column}

\item{info.default}{Default value to fill newly created info column, should data.set lack a column equivalent to info}
}
\value{
A data.table of fitted parameter values for each chemical.
}
\description{
Fits parameters of a specified model to concentration-time data given in
data.set
}
\author{
Caroline Ring, John Wambaugh
}
