% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all.R
\name{fit_all}
\alias{fit_all}
\title{Main fitting function}
\usage{
fit_all(data.set, model = "1compartment", modelfun = "analytic")
}
\arguments{
\item{data.set}{A table of concentration-time data. Preferably
\code{pkdataset_nheerlcleaned} or \code{pkdataset_nheerlorig}.}

\item{model}{Which general model should be fit for each chemical. Presently,
only "1compartment" and "2compartment" are implemented.}

\item{modelfun}{Either "analytic" or "full" -- whether to fit using the
analytic solution to the model, or the full ODE model. Presently,
"analytic" is recommended (because the analytic solution is exact and much
faster).}
}
\value{
A data.table of fitted parameter values for each chemical.
}
\description{
Fits parameters of a specified model to concentration-time data given in
data.set
}

