% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lower_bounds.R
\name{get_lower_bounds}
\alias{get_lower_bounds}
\title{Get lower bounds for fitting model parameters}
\usage{
get_lower_bounds(par_DF = NULL, model, fitdata, suppress.messages = FALSE)
}
\arguments{
\item{par_DF}{Optional: A data.frame as produced by [get_opt_params], with a
character variable [param_name] containing parameter names, and a logical
variable [optimize_param] containing TRUE if parameter is to be fitted, and
FALSE if parameter is to be held constant. Any other variables will be
ignored. Default NULL, in which case it will be determined by calling
[get_opt_params].}

\item{model}{The name of the model whose parameters are to be estimated.}

\item{fitdata}{A data.frame: the concentration-time-dose data to be used for
fitting.}
}
\value{
A data.frame: `parDF` with additional variables `lower_bound`
  (numeric, containing the lower bound for each parameter) and
  `lower_bound_msg` (character, containing a brief message explaining how the
  lower-bound value was calculated).
}
\description{
For a set of model parameters, get the lower bounds for the optimizer (if
parameter is to be fitted).
}
\author{
Caroline Ring, John Wambaugh, Mitchell Teague
}
