% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opt_params.R
\name{get_opt_params}
\alias{get_opt_params}
\title{Determine which parameters to fit (optimize)}
\usage{
get_opt_params(
  model,
  fitdata,
  param_names = NULL,
  pool_sigma = FALSE,
  suppress.messages = FALSE
)
}
\arguments{
\item{model}{Character: the name of the model to parameterize.}

\item{fitdata}{Data.frame: The set of concentration-dose-time data to be used
to fit the model parameters. Requires a variable named "Route" which
contains the dosing route, either "po" (oral dosing) or "iv" (intravenous
dosing).}

\item{param_names}{Optional: A character vector of parameter names needed by
the model. Default NULL to automatically determine these from `model` by
calling [get_model_paramnames].}

\item{pool_sigma}{Logical: Whether to pool all data (estimate only one error
standard deviation) or not (estimate separate error standard deviations for
each reference). Default FALSE to estimate separate error SDs for each
reference. (If `fitdata` only includes one reference, `pool_sigma` will have
no effect.)}
}
\value{
A `data.frame` with the following variables:
- `param_name` Character: the name of each model parameter.
- `optimize_param` Logical: Whether the parameter is to be
optimized/estimated from the data (TRUE), or not (FALSE). For example, if
`model` is '1compartment' and `fitdata` does not contain both oral and IV
data, then the parameter "Fgutabs" cannot be estimated from the data, so
`optimize_param` for "Fgutabs" will be FALSE.
- `use_param` Logical: Whether the parameter gets used in the model at all
(TRUE) or not (FALSE). For example, if `model` is '1compartment' and
`fitdata` contains no oral data at all, then for parameter "Fgutabs",
`use_param` will be FALSE, because the 1-compartment model function for
IV-only data will not use the value for "Fgutabs".

 By default, `use_param` and `opt_param` will be the same -- for each
 parameter, both `TRUE` or both `FALSE`. However, both variables are retained
 to allow for the possibility of holding one or more parameters constant and
 optimizing the rest (which would be accomplished by setting
`optimize_param  = FALSE` but keeping `use_param = TRUE` for the parameters
to be held constant).
}
\description{
Determine whether to fit (optimize) each model parameter from the data or
hold it constant
}
\details{
All model parameters will be fitted (optimized) with the following
exceptions:

For 1-compartment and 2-compartment models:

If no oral dosing data are available, then the parameter "kgutabs" will not
be estimated from the data. Instead it will be held constant at its
"starting" value while other parameters are estimated.

If data do not include both oral dosing and IV dosing, then the parameter
"Fgutabs" will not be estimated from the data. Instead it will be held
constant at its "starting" value while other parameters are estimated.

If non-NULL `param_names` do not match the result of
`get_model_parameters(param_names)`, this function will throw a warning, but
will otherwise proceed using `param_names`.
}
