% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_2comp.R
\name{get_params_2comp}
\alias{get_params_2comp}
\title{Get parameters for 2-compartment model}
\usage{
get_params_2comp(data)
}
\arguments{
\item{data}{The data set to be fitted, with harmonized variable names (e.g. the result of [preprocess_data()]).}
}
\value{
A `data.frame`with the following variables:
- `param_names`: Names of the model parameters
- `param_units`: Units of the model parameters
- `opt_params`: TRUE if each parameter is to be estimated from the data; FALSE otherrwise
- `use_params`: TRUE if each parameter is to be used in evaluating the model; FALSE otherwise
}
\description{
Get parameters for 2-compartment model and determine whether each is to be
estimated from the data
}
\details{
The full set of model parameters for the 1-compartment model includes `V1`,
`kelim`, `k12`, `k21`, `kgutabs`, and `Fgutabs`.

## IV data, no oral data

If IV dosing data are available, but no oral dosing data are available, then
only the parameters `V1`, `kelim`, `k12`, and `k21` will be estimated from
the data. The parameters `kgutabs` and `Fgutabs` cannot be estimated from IV
data alone.

## Oral data, no IV data

If oral dosing data are available, but no IV dosing data are available, then
the parameters `kelim`, `k12`, `k21`, and `kgutabs` will be estimated from
the data. However, the parameters `Fgutabs` and `V1` cannot be identified
separately. From oral data alone, only the ratio `Fgutabs/V1` can be
identified. This ratio is represented by a single parameter named
`Fgutabs_V1`. `Fgutabs` and `V1` will not be optimized, but `Fgutabs_V1` will
be optimized, along with `kelim`, `k12`, `k21`, and `kgutabs`.

## Oral data and IV data

If both oral and IV dosing data are available, then `V1`, `kelim`, `k12`,
`k21`, `kgutabs`, and `Fgutabs` will all be estimated from the data.

# Blood and plasma data

If both blood and plasma data are available, then `Rblood2plasma` will be estimated from the data.

# Blood data, no plasma data

If blood data but no plasma data are available, then `Rblood2plasma` will be used in model evaluation, but held constant at 1, not estimated from the data.

# Plasma data, no blood data

If plasma data but no blood data are available, then `Rblood2plasma` will neither be estimated nor be used in model evaluation at all.
}
\author{
Caroline Ring
}
