% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upper_bounds.R
\name{get_upper_bounds}
\alias{get_upper_bounds}
\title{Get upper bounds for fitting model parameters}
\usage{
get_upper_bounds(
  fitdata,
  model,
  par_DF = NULL,
  digits = 3,
  scientific = -2,
  suppress.messages = FALSE
)
}
\arguments{
\item{fitdata}{A data.frame: the concentration-time-dose data to be used for
fitting.}

\item{model}{The name of the model whose parameters are to be estimated.}

\item{par_DF}{Optional: A data.frame as produced by [get_opt_params], with a
character variable [param_name] containing parameter names, and a logical
variable [optimize_param] containing TRUE if parameter is to be fitted, and
FALSE if parameter is to be held constant. Any other variables will be
ignored. Default NULL, in which case it will be determined by calling
[get_opt_params].}
}
\value{
A data.frame: `parDF` with additional variables `upper_bound`
  (numeric, containing the upper bound for each parameter) and
  `upper_bound_msg` (character, containing a brief message explaining how the
  upper-bound value was calculated).
}
\description{
For a set of model parameters, get the upper bounds for the optimizer (if
parameter is to be fitted).
}
\author{
Caroline Ring, John Wambaugh, Mitchell Teague
}
