% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log-likelihood}
\usage{
log_likelihood(
  params,
  const_params = NULL,
  DF,
  modelfun = "analytic",
  model,
  fit_conc_dose = TRUE,
  force_finite = FALSE,
  negative = TRUE
)
}
\arguments{
\item{params}{A named vector of log-scaled parameter values. When this
function is the objective function for a numerical optimizer, these are the
parameters to be optimized.}

\item{const_params}{A named vector of additional log-scaled parameter values.
When this function is the objective function for a numerical optimizer,
these are additional model parameters whose value is to be held constant
while the other parameters are optimized. Default NULL (meaning that all
model parameters are supplied in `params`). (If you are calling this
function directly, you probably want to leave `const_params = NULL` and
just supply all model parameters in `params`.)}

\item{DF}{A `data.frame` of concentration-time data, for example as produced
by [preprocess_data()]. It must contain variables named `Study`
(representing the study ID for each observation), `Time` (representing
observation time points in hours), `Dose` (representing administered dose
in mg/kg), `iv` (logical TRUE/FALSE, representing whether the administered
dose was IV), `Value` (representing observed concentrations in mg/L), `LOQ`
(numeric, representing limit of quantification of each observed
concentration  in mg/L), and `N_Subjects` (representing the number of
subjects included in the concentration measurement). If any `N_Subjects` >
1, then `DF` must also include a variable `Value_SD`, representing the
sample standard deviation for multiple-subject observations (where `Value`
represents the sample mean of multiple measurements).}

\item{modelfun}{Character: whether to use the analytic TK model solution or
solve the full ODE model numerically. "analytic" to use the analytic model
solution, "full" to use the full ODE model. Default is "analytic".}

\item{model}{Character: The model to fit. Currently, only "flat",
"1compartment" or "2compartment" models are implemented.]}

\item{fit_conc_dose}{Logical: Whether to fit dose-normalized concentrations
(TRUE) or non-dose-normalized concentrations (FALSE). Default TRUE.}

\item{force_finite}{Logical: Whether to force the function to return a finite
log-likelihood (e.g., as required by [optimx::optimx()] with method
'L-BFGS-B'.) Default FALSE, allowing the function to return -Inf for
infinitely-unlikely parameter combinations. When `force_finite == TRUE`,
the function will replace -Inf with -999999.}

\item{negative}{Logical: Whether to return the *negative* log-likelihood
(i.e., the log-likelihood multiplied by negative 1). Default TRUE, to
return the negative log-likelihood. This option is useful when treating the
log-likelihood as an objective function to be minimized by an optimization
algorithm.}
}
\value{
A log-likelihood value for the data given the parameter values in
  params
}
\description{
The log-likelihood function (probability of data given model parameters).
}
\details{
The log-likelihood is formulated by assuming that residuals (model-predicted
concentrations minus observed concentrations) are independent and identically
distributed within a given combination of chemical, species, and study,
obeying a zero-mean normal distribution with a chemical-, species-, and
study-specific variance.

# Log-likelihood equations

For chemical-species combination \eqn{i} and study \eqn{j}, define the
following quantities.

\eqn{y_{ijk}} is the \eqn{k^{th}} observation of concentration, corresponding
to dose \eqn{d_{ijk}} and time \eqn{t_{ijk}}. Each observation has a
corresponding LOQ, \eqn{\textrm{LOQ}_{ijk}}.

For multiple-subject observations, \eqn{y_{ijk}} is the \eqn{k^{th}} *sample
mean* observed concentration for chemical-species combination \eqn{i} and study \eqn{j},
corresponding to dose \eqn{d_{ijk}} and time \eqn{t_{ijk}}. It represents the
mean of \eqn{n_{ijk}} individual measurements. It has a corresponding sample
standard deviation, \eqn{s_{ijk}}.

\eqn{\bar{theta}_i} represents the vector of model parameters supplied in
argument `params` for chemical-species combination \eqn{i}.

\eqn{\mu_{ijk}} is the model-predicted concentration for dose \eqn{d_{ijk}}
and time \eqn{t_{ijk}}. If \eqn{f(d, t; \bar{ \theta})} is the model function
evaluated at dose \eqn{d} and time \eqn{t}, with parameter vector
\eqn{\bar{\theta}}, then

 \deqn{\mu_{ijk} = f \left(d_{ijk}, t_{ijk};
\bar{\theta}_i \right)}

\eqn{\sigma_{ij}^2} is the study- and chemical-specific residual
variance. (It is a hyperparameter.)

## Single-subject observations above limit of quantification (detects)

This is the normal probability density function evaluated at the observed
concentration, as implemented in [stats::dnorm()].

\deqn{LL_{ijk} = \log \left( \frac{1}{\sqrt{\sigma_{ij} 2 \pi}} \exp \left[
\frac{-1}{2} \left( \frac{y_{ijk} - \mu_{ijk}}{\sigma_{ij}} \right)^2
\right]   \right) }


## Single-subject observations below limit of quantification (non-detects)

This is the normal cumulative density function evaluated at the LOQ, as
implemented in [stats::pnorm()]. It is the total probability of observing a
concentration anywhere between 0 and the LOQ.

\deqn{LL_{ijk} =  \log \left( \frac{1}{2} \left[ 1 + \textrm{erf} \left(
\frac{\textrm{LOQ}_{ijk} - \mu_{ijk}}{\sigma_{ij} \sqrt{2}} \right) \right]  \right) }

## Multiple-subject observations above limit of quantification

This is the joint log-likelihood across the multiple subjects included in one
observation, re-expressed in terms of the sample mean, sample SD, and number
of subjects for that observation. It is implemented in [dnorm_summary()].

\deqn{LL_{ijk} = n_{ijk} * \log \left[
 \frac{1}{\sigma_{ij} \sqrt{2 \pi}} +
  \frac{-1}{2 \sigma_{ij}^2}  \left(
   \left( n_{ijk} - 1 \right)
     s_{ijk}^2 + n_{ijk} \left( y_{ijk} - \mu_{ijk} \right)^2
      \right)
       \right]}

## Multiple-subject observations below limit of quantification

This case is not implemented. If sample mean concentration is reported below
LOQ, then it is unclear what individual observed concentrations are
represented, and how they were combined to produce the summary data in terms
of sample mean and sample SD. Were all individual observations below LOQ? Or
were below-LOQ observations replaced with 0, LOQ/2, etc. before sample mean
and sample SD were computed? If the sample mean is reported below LOQ, what
LOQ is reported? Did individual observations all have the same LOQ, or is an
average or median LOQ being used?  It is impossible to formulate the
log-likelihood without knowing the answers to these questions. Therefore,
multiple-subject observations below LOQ are excluded from analysis (they are
removed in [preprocess_data()]).

# Joint log-likelihood for a chemical and species

The joint log-likelihood for a chemical and species is simply the sum of
log-likelihoods across observations and studys.

\deqn{LL_{i} = \sum_{j=1}^{J_i} \sum_{k=1}^{K_{ij}} LL_{ijk}}

This is the overall probability of the observed data, given the model and
parameters.
}
