% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conctime.R
\name{plot_conctime}
\alias{plot_conctime}
\title{plot_conctime}
\usage{
plot_conctime(
  PK.fit.table,
  data.set,
  model,
  mean.type = "Fitted geometric mean",
  fit.factors = c("Compound", "Data.Analyzed", "Dose.nominal", "Route", "LOQ"),
  plot.split.factor = "Route",
  color.factor = "Data.Analyzed",
  plot.httk.pred = F,
  omit.zero = T,
  plot.loq = T,
  fit.plot.points = 200,
  data.dose.col = "dose_level_normalized",
  data.compound.col = "preferred_name",
  data.cas.col = "dsstox_casrn",
  data.source.col = "fk_study_id",
  data.route.col = "administration_route_normalized",
  data.reference.col = "fk_reference_document_id",
  data.media.col = "conc_medium_normalized",
  data.loq.col = "calc_loq",
  data.value.col = "conc",
  data.time.col = "time_hr"
)
}
\arguments{
\item{PK.fit.table}{Table of parameter estimates from \code{\link{fit_all}}}

\item{data.set}{A data.table of concentration vs. time data for a given
chemical}

\item{model}{Analytic model to evaluate. Currently only "1compartment" or
"2compartment" are implemented.}

\item{mean.type}{(Defaults to fitted geometric mean)}

\item{fit.factors}{Add a trend-line using the fit for each combination of
the factors listed in this character vector. (defaults to Compound,
Dose.nominal, Data.Analyzed, and Route)}

\item{plot.split.factor}{For each compound, make a separate plot for each
unique value of this factor (defaults to Route) #param shape.factor Indicate
this factor by shape in the plots (defaulta to Data.Analyzed)}

\item{color.factor}{Indicate this factor by color in the plots (defaults to
Data.Analyzed')}

\item{plot.httk.pred}{placeholder}

\item{omit.zero}{Drop zero dose values from plots (default = TRUE)}

\item{plot.loq}{placeholder}

\item{fit.plot.points}{placeholder}
}
\value{
None
}
\description{
Creates a PDF of the fits from \code{\link{fit_all}} and data together
}
\author{
Caroline Ring, John Wambaugh
}
