% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_preprocess_data.R
\name{preprocess_data.pk}
\alias{preprocess_data.pk}
\title{Pre-process data}
\usage{
\method{preprocess_data}{pk}(obj)
}
\arguments{
\item{obj}{A `pk` object}
}
\value{
The same `pk` object, with added elements `data` (containing the
  cleaned, gap-filled data) and `data_info` (containing summary information
  about the data, e.g. number of observations by route, media,
  detect/nondetect; empirical tmax, time of peak concentration for oral data;
  number of observations before and after empirical tmax)
}
\description{
Pre-process data for a `pk` object
}
\details{
Data pre-processing for an object `obj` includes the following steps, in order:
1. Coerce data to class `data.frame` (if it is not already)
2. Rename variables to harmonized "`invivopkfit` aesthetic" variable names, using `obj$mapping`
3. Check that the data includes only one chemical and one species.
4. Check that the data includes only routes in `obj$data_settings$routes_keep` and media in `obj$data_settings$media_keep`
5. Check that the data includes only one unit for concentration, one unit for time, and one unit for dose.
6. Coerce Value, Value_SD, LOQ, Dose, and Time to numeric, if they are not already.
7. Coerce Species, Route, and Media to lowercase.
8. Replace any negative Value, Value_SD, Dose, or Time with `NA`
9. Replace any negative
}
\author{
Caroline Ring
}
