% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process concentration vs. time data for analysis}
\usage{
preprocess_data(
  data.set,
  names_list = list(Compound_Dosed = "studies.test_substance_name_original", DTXSID_Dosed
    = "chemicals_dosed.dsstox_substance_id", CAS_Dosed = "chemicals_dosed.dsstox_casrn",
    Compound_Analyzed = "series.analyte_name_original", DTXSID_Analyzed =
    "chemicals_analyzed.dsstox_substance_id", CAS_Analyzed =
    "chemicals_analyzed.dsstox_casrn", Reference = "documents_reference.id", Extraction =
    "documents_extraction.id", Species = "subjects.species", Weight =
    "subjects.weight_kg", Weight.Units = NULL, Dose = "studies.dose_level_normalized", 
 
       Dose.Units = NULL, Time = "conc_time_values.time_hr", Time.Units = NULL, Media =
    "series.conc_medium_normalized", Value = "conc_time_values.conc", Value.Units = NULL,
    Route = "studies.administration_route_normalized", LOQ = "series.loq", Subject =
    "subjects.id", N_Subjects = "series.n_subjects_in_series", Value_SD =
    "conc_time_values.conc_sd", Study_ID = "studies.id", Series_ID = "series.id"),
  defaults_list = list(Weight.Units = "kg", Dose.Units = "mg/kg", Time.Units = "hours",
    Value.Units = "mg/L"),
  ratio_conc_to_dose = 1,
  calc_loq_factor = 0.45,
  routes_keep = c("po", "iv"),
  media_keep = c("blood", "plasma"),
  impute_loq = TRUE,
  impute_sd = TRUE,
  study_def = c("DTXSID", "Species", "Reference", "Route", "Media"),
  suppress.messages = FALSE
)
}
\arguments{
\item{data.set}{A `data.frame` of concentration-time data. Preferably
`pkdataset_nheerlcleaned` or `pkdataset_nheerlorig`.}

\item{names_list}{As for [rename_columns()]: A named list where the names are
the new variable names, and the values are the old variable names. If a
value is NULL, then there is no old variable corresponding to that new
variable. A new variable will be added, with default value as given in
`defaults_list`; if no default is given in `defaults_list`, the new
variable will be filled with `NA_character`.}

\item{defaults_list}{As for [rename_columns()]: A named list where the names
are the new variable names, and the values are the default values to fill
for each new variable, if the corresponding old variable name is NULL in
`names_list`.}

\item{ratio_conc_to_dose}{Ratio between the mass units used to report the
concentration data and the mass units used to report the dose. Default 1.
For example, concentration reported in ug/L and dose reported in mg/kg/day
would require `ratio_conc_to_dose = 0.001`, because 1 ug/1 mg = 1e-6 g /
1e-3 g = 0.001.}

\item{routes_keep}{List of routes to retain. Default `c("po", "iv")` to
retain only oral and IV administration data.}

\item{media_keep}{List of media to retain. Default `c("blood", "plasma")` to
retain only concentrations in blood and plasma.}

\item{impute_loq}{Logical: TRUE to impute values for missing LOQs; FALSE to
leave them alone.}

\item{impute_sd}{Logical: TRUE to impute values for missing sample SDs for
multi-subject observations; FALSE to leave them alone}

\item{study_def}{A character vector specifying the variables (new names)
whose unique combinations define individual "studies" (where each "study"
will have its own error SD in the fitting process). Default is `c("DTXSID",
"Species", "Reference", "Route", "Media")`.}

\item{suppress.messages}{Logical: Whether to suppress verbose messages.
Default FALSE, to be verbose.}

\item{model}{Which general model should be fit for each chemical. Presently,
only "1compartment" and "2compartment" are implemented.}

\item{modelfun}{Either "analytic" or "full" -- whether to fit using the
analytic solution to the model, or the full ODE model. Presently,
"analytic" is recommended (because the analytic solution is exact and much
faster).}
}
\value{
A `data.table` containing the cleaned, harmonized data, ready for
  model fitting. This data.table contains the following variables:
  \describe{
  \item{`Compound`}{Chemical compound name}
  \item{`DTXSID`}{DSSTox Substance ID,}
  \item{`CAS`}{Chemical CASRN}
  \item{`Reference`}{Study reference document ID}
  \item{`Species`}{Species.}
  \item{`Weight`}{Body weight}
  \item{`Weight_Units`}{Units of body weight`}
  \item{`Dose`}{Dose for each observation}
  \item{`Time`}{Time of each observation}
  \item{`Time.Units`}{Units of times in `Time`}
  \item{`Media`}{Medium in which concentrations were measured, e.g. "blood"
  or "plasma".}
  \item{`Value`}{Concentration values of each observation}
  \item{`Units`}{Units of concentration values in `Value`}
  \item{`Route`}{Route of dose administration: either "iv" (intravenous) or
  "po" (oral). Observations with any other routes are removed.}
  \item{`Extraction`}{Study extraction document ID.}
  \item{`LOQ`}{Limit of quantitation for concentrations}
  \item{`Subject`}{Individual subject identifier, if available. }
  \item{`iv`}{Logical TRUE/FALSE flag indicating whether `Route == "iv"`.}
  \item{`Time.Days`}{Time of each observation, converted to units of days.}
  \item{`N.Obs.Ref`}{The number of observations remaining for each unique
  combination of `Reference`, `DTXSID`, and `Species`, after the removal
  steps described in Details.}
  }

  new_name = old_name
}
\description{
Clean data to set up for PK model fitting.
}
\details{
# Preprocessing steps

This function does the following things in the following order, and is
verbose about it unless told otherwise (by setting `suppress.messages =
FALSE`):

- Renames columns to the standard names that invivoPKfit uses internally, and
adds any missing columns. See [rename_columns()].
- Converts concentrations to numeric, if they are not already.
- Converts doses to numeric, if they are not already.
- Converts times to numeric, if they are not already.
- Converts reference ID to character, if not already.
- Converts extraction source ID to character, if not already.
- If reference ID is NA, sets it to be the same as extraction source ID. (By
default in CvTdb, reference ID is only set if it is different from
extraction-source ID. For example, if data were extracted from a figure in a
meta-analysis paper that republished data from multiple other publications,
the extraction-source ID would be for the meta-analysis paper, but the
reference ID would be for the original publication.)
- Removes all observations where `DTXSID_Dosed` and `DTXSID_Analyzed` are not
the same (i.e., keeps only data where the parent chemical is what was
monitored). Creates new columns `Compound`, `DTXSID`, and `CAS` identifying
the chemical both dosed and analyzed.
- Harmonizes routes recorded as "oral" to "po" and "intravenous" to "iv".
- Removes all observations that have routes other than those listed in `routes_keep`.
- Removes all observations in media other than those listed in `media_keep`.
- Adds variable `iv`: a TRUE/FALSE flag, for whether route is "iv" or not.
- Converts `Compound` and `Species` to lower-case.
- Multiplies concentrations by `ratio_conc_to_dose` (*i.e.*, by the ratio
between the mass units for concentration and the mass units for dose).
- For any concentrations reported as 0, substitute NA.
 - Imputes LOQ for any observations missing it, using [estimate_loq()]. LOQ
is imputed for each unique combination of `Reference`, `DTXSID`, `Species`,
and `Media` as `calc_loq_factor` times the minimum *detected* `Value`.s
 - Substitutes NA for any concentration observations below LOQ (these are non-detects).
 - Removes any remaining observations where both `Value` and `LOQ` are NA.
For example, this situation occurs for reference/chemical/species/media
combinations where no LOQ was reported, and all concentrations were reported
as NA, so that no LOQ could be imputed.
- Removes any observations with NA `Time`.
- Adds a variable `Time.Days` where `Time` (in hours) is converted to time in days.
}
