% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_data_settings.R
\name{settings_data}
\alias{settings_data}
\title{Data preprocessing settings}
\usage{
settings_data(
  routes_keep = c("oral", "iv"),
  media_keep = c("blood", "plasma"),
  impute_loq = TRUE,
  loq_group = vars(Chemical, Species, Reference, Media),
  calc_loq_factor = 0.45,
  impute_sd = TRUE,
  sd_group = vars(Chemical, Species, Reference, Media),
  suppress.messages = FALSE,
  ...
)
}
\arguments{
\item{suppress.messages}{}
}
\value{
An object of class `pk_data_settings`.
}
\description{
Data preprocessing settings
}
\author{
Caroline Ring
}
